### 2019-02-28 #1289 @whedon
[PRE REVIEW]: PyBIDS: Python tools for BIDS datasets
**Submitting author:** @tyarkoni (<a href="http://orcid.org/0000-0002-6558-5113">Tal Yarkoni</a>)
**Repository:** <a href="https://github.com/bids-standard/pybids" target ="_blank">https://github.com/bids-standard/pybids</a>
**Version:** 0.8.0
