### 2019-02-28 #1289 @danielskatz
👋 @tyarkoni - your paper.md authors and affiliations aren't formatted right - see https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for an example of how this needs to be.
