### 2019-02-28 2eab0d3
fixes to paper.md
