+++ paper/paper.md
@@ -5,108 +5,151 @@ tags:
   - BIDS
   - neuroimaging
 authors:
-  - affiliation: University of Texas at Austin
+  - affiliation: 1
     name: Yarkoni, Tal
     orcid: 0000-0002-6558-5113
-  - affiliation: Stanford University
+  - affiliation: 2
     name: Markiewicz, Christopher J.
     orcid: 0000-0002-6533-164X
-  - affiliation: University of Texas at Austin
+  - affiliation: 1
     name: de la Vega, Alejandro
     orcid: 0000-0001-9062-3778
-  - affiliation: Department of Psychology, Stanford University
+  - affiliation: 2
     name: Gorgolewski, Krzysztof J.
     orcid: 0000-0003-3321-7583
-  - affiliation: Florida International University
+  - affiliation: 3
     name: Salo, Taylor
     orcid: 0000-0001-9813-3167
-  - affiliation: Dartmouth College
+  - affiliation: 4
     name: Halchenko, Yaroslav O.
     orcid: 0000-0003-3456-2493
-  - affiliation: University of Texas at Austin
+  - affiliation: 1
     name: McNamara, Quinten
-  - affiliation: University of Oregon
+  - affiliation: 5
     name: DeStasio, Krista
     orcid: 0000-0002-3959-9060
-  - affiliation: McGill University
+  - affiliation: 6
     name: Poline, Jean-Baptiste
     orcid: 0000-0002-9794-749X
-  - affiliation: Institute for Problems of Information Transmission
+  - affiliation: 7
     name: Petrov, Dmitry
-  - affiliation: Concordia University
+  - affiliation: 8
     name: Hayot-Sasson, Valérie
     orcid: 0000-0002-4830-4535
-  - affiliation: National Institute of Mental Health
+  - affiliation: 9
     name: Nielson, Dylan M.
     orcid: 0000-0003-4613-6643
-  - affiliation: MRC Cognition and Brain Sciences Unit
+  - affiliation: 10
     name: Carlin, Johan
     orcid: 0000-0003-0933-1239
-  - affiliation: Montreal Neurological Institute and Hospital
+  - affiliation: 11
     name: Kiar, Gregory
     orcid: 0000-0001-8915-496X
-  - affiliation: Alan Turing Institute; Department of Psychiatry, University of Cambridge
+  - affiliation: 12
     name: Whitaker, Kirstie
     orcid: 0000-0001-8498-4059
-  - affiliation: Montreal Neurological Institute and Hospital
+  - affiliation: 11
     name: DuPre, Elizabeth
     orcid: 0000-0003-1358-196X
-  - affiliation: Psychoinformatics Lab, Institute of Psychology, Otto-von-Guericke University
-      Magdeburg
+  - affiliation: 13
     name: Wagner, Adina
     orcid: 0000-0003-2917-3450
   - name: Ivanov, Alexander
-  - affiliation: CorticoMetrics LLC
+  - affiliation: 14
     name: Tirrell, Lee S.
     orcid: 0000-0001-9393-8361
-  - affiliation: Télécom ParisTech, France
+  - affiliation: 15
     name: Jas, Mainak
     orcid: 0000-0002-3199-9027
-  - affiliation: Otto-von-Guericke-University Magdeburg, Germany
+  - affiliation: 13
     name: Hanke, Michael
     orcid: 0000-0001-6398-6370
-  - affiliation: Stanford University
+  - affiliation: 2
     name: Poldrack, Russell
     orcid: 0000-0001-6755-0259
-  - affiliation: Department of Psychology, Stanford University
+  - affiliation: 2
     name: Esteban, Oscar
     orcid: 0000-0001-8435-6191
-  - affiliation: Center for Adaptive Rationality, Max Planck Institute for Human Development,
[…]
