### 2019-02-28 #1289 @tyarkoni
Re: editor, I would have requested @arokem because of his Python expertise, but he's a contributor to the repository, so it seemed like a conflict of interest.

There's a PR that I *think* fixes the issues. Let's see...
