### 2019-02-28 #1289 @danielskatz
Whedon commands need to be new comments
