### 2019-03-01 #1289 @grlee77
I am interested in volunteering as a reviewer for this submission. I have MRI neuroimaging and python expertise.
I do not have any conflicts to report (I have contributed to one of the project's dependencies, NiBabel, but have not had any involvement with BIDS or pyBIDS).
