### 2019-03-01 29b9df5
Do not convert to Dense if variable is Sparse
