### 2019-03-02 ffdaa01
Set resample_frames to uniform 10hz if Sparse, return Dense
