### 2019-03-02 857cd0f
get_duration on var
