### 2019-03-04 #1294 @tyarkoni
I adapted it from the repo's [.zenodo.json](https://github.com/bids-standard/pybids/blob/master/.zenodo.json) file. Convention for that seems to be [LastName], [FirstName].
