### 2019-03-04 40c5dec
Default to 1 Hz
