### 2019-03-04 625d67d
Return events at 10Hz, enable adapative oversampling
