### 2019-03-04 b4881c8
Take min of duration, and do future division
