### 2019-03-04 129a30b
Simplify adapative oversampling code
