### 2019-03-05 44825e2
Update bids/analysis/transformations/compute.py
Co-Authored-By: adelavega <aleph4@gmail.com>
