### 2019-03-05 49547ea
Lint convolve function
