### 2019-03-05 258c26e
Add min_interval comment
