### 2019-03-05 38f6cc4
Compute min_interval based on unique values
