### 2019-03-05 46f638e
Add test_convolve, with mocking for testing oversampling factor
