### 2019-03-05 2e4d736
ENH: do not hardcode sampling rate(s) and expected values - use/compute
