### 2019-03-05 24646c8
Double oversampling, (pre ceil rounding)
