### 2019-03-05 101956c
OPT(TST): select only 2 subjects for testing, hardcode even less numbers
I thought also to decrease number of runs, but then test_split is not behaving,
so decided to stay at this.  Now tests here would run for 25-30 sec in contrast to 150-180
on my laptop
