### 2019-03-05 d3e50ee
Check that there are at least two events to compute min diff
