### 2019-03-05 9a1c357
OPT(TST): cache ds005 collection and return a clone upon invocation
Creating a collection apparently is more expensive.
Clone() is needed since some tests do modification inplace. If they did not
it was possible to declare this fixture as (scope="module"). So we need to return
a clone each time.  This causes full run time of 9 sec vs 30 sec.
