### 2019-03-06 916c8a3
Use same formula as DenseRunVariable._build_entity_index to calculate frame_times
