### 2019-03-11 d5f6b93
ENH: Initial framework for sparse acquisition
