### 2019-03-11 0729028
ENH: Add method to choose sampling rate
