### 2019-03-12 bids-standard/pybids#415 @grlee77
minor demo/documentation typos
Cell 9 of the demo IPython notebook states
```Python
# Ask get() to return the ids of subjects that have T1w files
layout.get(return_type='id', target='subject')
```
but I think this should be specifying the suffic argument to do what is described in the comment?
```Python
# Ask get() to return the ids of subjects that have T1w files
layout.get(suffix='T1w', return_type='id', target='subject')
```

Also in section 3.2 of the docs, the text refers to `model.json` but the example code uses `model1.json` instead.
