### 2019-03-12 bids-standard/pybids#416 @grlee77
make dependencies easier to find
In reviewing for JOSS, one of the documentation criteria is:

**Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Your package installed fine and included the necessary dependencies automatically via pip, but I do not see the dependencies themselves listed anywhere on the installation section of the docs or the top-level README.

I was able to find them by inspecting the source, but it would be helpful to have them explicitly stated so a user can know what to expect. For example, a `conda` user may prefer to install most dependencies via `conda` rather than have `pip` automatically install them.
