### 2019-03-12 c2db61f
BF: 0.0.7 version of duecredit stub, addresses import issue check on py3
