### 2019-03-13 #1294 @grlee77
It looks like the approach here was to include all contributors to the repository. I  just had a question about whether all coauthors have already actively confirmed their desire to be included?

@cmadan:  If past contributors have not been notified and explicitly agreed to be included in the publication then it seems they would need to be removed from the list?  The JOSS author guidelines state

"The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work."
