### 2019-03-13 #1294 @grlee77
The following are not specific to review checklist above, but more about two things that were not entirely clear to me as a new user when going through the examples in the documentation.

1.) There was an example using the `path_patterns` argument to the `build_path` method of a `BIDSLayout`. However, the API docs do not seem to explicitly describe the expected syntax for this argument.

2.)  In the documentation section on generating reports, it is mentioned that the reports are returned as a `collections.Counter`, but it isn't mentioned why this would be useful format or much detail on how it should typically be used. The example shown just uses

```Python
main_report = counter.most_common()[0][0]
```
but does not comment on why the indexing [0][0] is needed or when one would typically want or need to use the other elements returned by `most_common`?
