### 2019-03-19 31b20a5
list dependencies in README; closes #416
