### 2019-03-19 db0b764
fix/update demo notebook; closes #415
