### 2019-03-19 5843494
add note about Python version/OS support
