### 2019-03-19 47b9549
add community guidelines section to README
