### 2019-03-19 1f62fd9
expand build_path docstring for path_patterns arg
