### 2019-03-19 #1294 @tyarkoni
@grlee77 I think I've now addressed all of the comments in this thread and on the pybids repo, pending one minor issue that I've asked @tsalo (who authored the code in question) to look at. The [joss-review](https://github.com/bids-standard/pybids/tree/joss-review) branch collects all changes. Please let me know if there's anything else. Thanks for your effort and excellent feedback!
