### 2019-03-19 #1294 @tyarkoni
I think adding that rationale to the docstring would work. Feel free to add to the `joss-review` branch (I added you as a collaborator).
