### 2019-03-19 #1294 @effigies
As another data point, there was a metadata gathering thread (https://github.com/bids-standard/pybids/pull/308) for Zenodo. There were a couple people who never responded, but none who requested not to be authors. I recognize Zenodo isn't JOSS, but if you're looking for active consent to be considered authors on the software (if not the paper), that thread is a worthwhile one.
