### 2019-03-21 4066cff
minor docstring changes
