### 2019-03-22 209f807
BF: use ordereddict because otherwise columns might not be correctly assigned
in the code .values() is accessed later on, and if order is random, would not
correspond to the actual columns values, causes "random effects".
Closes #423
