### 2019-03-23 #1294 @tyarkoni
Okay, see the request [here](https://github.com/bids-standard/pybids/pull/395#issuecomment-475882753). I haven't indicated exactly what we'll take a non-response to mean, in the hopes of pushing people to respond, but let's see what happens.
