### 2019-03-26 0ee00ce
Updated keyword-change with new keyword 'acquisition'
