### 2019-03-27 586f22f
clarify necessity of dataset_description.json for derivs and raise warning
