### 2019-03-28 a050b95
first partially working solution
