+++ bids/layout/models.py
@@ -1,0 +1,291 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.ext.associationproxy import association_proxy
+from sqlalchemy.orm.collections import attribute_mapped_collection
+from sqlalchemy import Column, Integer, String, Boolean, ForeignKey
+from sqlalchemy.orm import reconstructor, relationship, backref
+import re
+import os
+import warnings
+from copy import deepcopy
+
+from .writing import build_path, write_contents_to_file
+from ..utils import listify, check_path_matches_patterns
+from ..config import get_option
+from ..external import six
+
+
+Base = declarative_base()
+
+
+class BIDSFile(Base):
+    __tablename__ = 'files'
+
+    path = Column(String, nullable=False, unique=True, primary_key=True)
+    filename = Column(String)
+    dirname = Column(String)
+    scope = Column(String)
+    entities = association_proxy("tags", "value")
+    # entities = association_proxy('tags', 'entities',
+    #     creator=lambda name, args: Tag(ent_name=name, entity=args[0],
+    #                                        value=args[1]))
+
+    def __init__(self, filename, scope='bids'):
+        self.path = filename
+        self.filename = os.path.basename(self.path)
+        self.dirname = os.path.dirname(self.path)
+        self.scope = scope
+
+    def _matches(self, entities=None, extensions=None, regex_search=False):
+        """
+        Checks whether the file matches all of the passed entities and
+        extensions.
+
+        Args:
+            entities (dict): A dictionary of entity names -> values to match.
+            extensions (str, list): One or more file extensions to allow.
+            regex_search (bool): Whether to require exact match (False) or
+                regex search (True) when comparing the query string to each
+                entity.
+        Returns:
+            True if _all_ entities and extensions match; False otherwise.
+        """
+        if extensions is not None:
+            extensions = map(re.escape, listify(extensions))
+            extensions = '(' + '|'.join(extensions) + ')$'
+            if re.search(extensions, self.filename) is None:
+                return False
+
+        if entities is None:
+            return True
+
+        for name, val in entities.items():
+
+            if (name not in self.entities) ^ (val is None):
+                return False
+
+            if val is None:
+                continue
+
+            def make_patt(x):
+                patt = str(x)
+                if not regex_search:
+                    patt = re.escape(patt)
+                if isinstance(x, (int, float)):
+                    # allow for leading zeros if a number was specified
+                    # regardless of regex_search
+                    patt = '0*' + patt
+                if not regex_search:
+                    patt = '^{}$'.format(patt)
+                return patt
+
+            ent_patts = [make_patt(x) for x in listify(val)]
+            patt = '|'.join(ent_patts)
+
+            if re.search(patt, str(self.entities[name])) is None:
+                return False
+
+        return True
+
+    def copy(self, path_patterns, symbolic_link=False, root=None,
+             conflicts='fail'):
+        """ Copy the contents of a file to a new location.
+        
+        Args:
+            path_patterns (list): List of patterns use to construct the new
+                filename. See build_path documentation for details.
+            symbolic_link (bool): If True, use a symbolic link to point to the
+                existing file. If False, creates a new file.
+            root (str): Optional path to prepend to the constructed filename.
+            conflicts (str): Defines the desired action when the output path
+                already exists. Must be one of:
+                    'fail': raises an exception
+                    'skip' does nothing
+                    'overwrite': overwrites the existing file
+                    'append': adds  a suffix to each file copy, starting with 1
+        """
+        new_filename = build_path(self.entities, path_patterns)
[…]
