+++ bids/layout/tests/test_models.py
@@ -1,0 +1,105 @@
+import os
+import pytest
+import bids
+# from bids.layout import BIDSLayout, parse_file_entities, add_config_paths
+from bids.layout.models import BIDSFile, Entity, Tag, Base
+# from bids.layout.core import BIDSFile, Entity, Config
+# from os.path import join, abspath, basename, dirname
+# from bids.tests import get_test_data_path
+# from bids.utils import natural_sort
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+
+@pytest.fixture(scope="module")
+def session():
+    engine = create_engine('sqlite://')
+    Base.metadata.create_all(engine)
+    Session = sessionmaker(bind=engine)
+    return Session()
+
+@pytest.fixture
+def sample_bidsfile(tmpdir):
+    testfile = 'sub-03_ses-2_task-rest_acq-fullbrain_run-2_bold.nii.gz'
+    fn = tmpdir.mkdir("tmp").join(testfile)
+    fn.write('###')
+    return BIDSFile(os.path.join(str(fn)))
+
+
+@pytest.fixture(scope='module')
+def subject_entity():
+    return Entity('subject', "[/\\\\]sub-([a-zA-Z0-9]+)", mandatory=False,
+               directory="{subject}", dtype='str')
+
+
+def test_entity_initialization():
+    e = Entity('avaricious', r'aardvark-(\d+)')
+    assert e.name == 'avaricious'
+    assert e.pattern == r'aardvark-(\d+)'
+    assert not e.mandatory
+    assert e.directory is None
+    assert e.files == {}
+
+
+def test_entity_init_all_args(subject_entity):
+    ent = subject_entity
+    assert ent.name == 'subject'
+    assert ent.pattern == "[/\\\\]sub-([a-zA-Z0-9]+)"
+    assert ent.mandatory == False
+    assert ent.directory == "{subject}"
+    assert ent.map_func is None
+    assert ent.kwargs == {'bleargh': True}
+
+
+def test_entity_init_with_bad_dtype():
+    with pytest.raises(ValueError) as exc:
+        ent = Entity('test', dtype='superfloat')
+        msg = exc.value.message
+        assert msg.startswith("Invalid dtype")
+
+
+def test_entity_deepcopy(subject_entity):
+    e = subject_entity
+    clone = copy.deepcopy(subject_entity)
+    for attr in ['name', 'pattern', 'mandatory', 'directory', 'map_func',
+                 'regex', 'kwargs']:
+        assert getattr(e, attr) == getattr(clone, attr)
+    assert e != clone
+
+
+def test_entity_matches(tmpdir):
+    filename = "aardvark-4-reporting-for-duty.txt"
+    tmpdir.mkdir("tmp").join(filename).write("###")
+    f = BIDSFile(os.path.join(str(tmpdir), filename))
+    e = Entity('avaricious', r'aardvark-(\d+)')
+    result = e.match_file(f)
+    assert result == '4'
+
+
+def test_entity_matches_with_map_func(sample_bidsfile):
+    bf = sample_bidsfile
+    e = Entity('test', map_func=lambda x: x.filename.split('-')[1])
+    assert e.match_file(bf) == '03_ses'
+
+
+def test_entity_unique_and_count():
+    e = Entity('prop', r'-(\d+)')
+    e.files = {
+        'test1-10.txt': '10',
+        'test2-7.txt': '7',
+        'test3-7.txt': '7'
+    }
+    assert sorted(e.unique()) == ['10', '7']
+    assert e.count() == 2
+    assert e.count(files=True) == 3
+
+
+def test_entity_add_file(sample_bidsfile, session):
[…]
