### 2019-03-28 e272021
fix dtype handling in Tag init
