### 2019-03-28 d05b1ba
load or create database on Layout init
