### 2019-03-28 835bd7c
make SQLAlchemy session globally accessible
