### 2019-03-28 c0ceb5e
temporarily replace core classes with SQLAlchemy models
