+++ bids/layout/core.py
@@ -8,6 +8,7 @@ import warnings
 from copy import deepcopy
 
 from .writing import build_path, write_contents_to_file
+from .models import BIDSFile, Entity, Tag
 from ..utils import listify, check_path_matches_patterns
 from ..config import get_option
 from ..external import six
@@ -58,260 +59,260 @@ class Config(object):
         return Config(**config)
 
 
-class Entity(object):
+# class Entity(object):
 
-    def __init__(self, name, pattern=None, mandatory=False, directory=None,
-                 map_func=None, dtype=None, **kwargs):
-        """
-        Represents a single entity defined in the JSON config.
+#     def __init__(self, name, pattern=None, mandatory=False, directory=None,
+#                  map_func=None, dtype=None, **kwargs):
+#         """
+#         Represents a single entity defined in the JSON config.
 
-        Args:
-            name (str): The name of the entity (e.g., 'subject', 'run', etc.)
-            pattern (str): A regex pattern used to match against file names.
-                Must define at least one group, and only the first group is
-                kept as the match.
-            mandatory (bool): If True, every File _must_ match this entity.
-            directory (str): Optional pattern defining a directory associated
-                with the entity.
-            map_func (callable): Optional callable used to extract the Entity's
-                value from the passed string (instead of trying to match on the
-                defined .pattern).
-            dtype (str): The optional data type of the Entity values. Must be
-                one of 'int', 'float', 'bool', or 'str'. If None, no type
-                enforcement will be attempted, which means the dtype of the
-                value may be unpredictable.
-            kwargs (dict): Additional keyword arguments.
-        """
-        if pattern is None and map_func is None:
-            raise ValueError("Invalid specification for Entity '%s'; no "
-                             "pattern or mapping function provided. Either the"
-                             " 'pattern' or the 'map_func' arguments must be "
-                             "set." % name)
-        self.name = name
-        self.pattern = pattern
-        self.mandatory = mandatory
-        self.directory = directory
-        self.map_func = map_func
-        self.kwargs = kwargs
+#         Args:
+#             name (str): The name of the entity (e.g., 'subject', 'run', etc.)
+#             pattern (str): A regex pattern used to match against file names.
+#                 Must define at least one group, and only the first group is
+#                 kept as the match.
+#             mandatory (bool): If True, every File _must_ match this entity.
+#             directory (str): Optional pattern defining a directory associated
+#                 with the entity.
+#             map_func (callable): Optional callable used to extract the Entity's
+#                 value from the passed string (instead of trying to match on the
+#                 defined .pattern).
+#             dtype (str): The optional data type of the Entity values. Must be
+#                 one of 'int', 'float', 'bool', or 'str'. If None, no type
+#                 enforcement will be attempted, which means the dtype of the
+#                 value may be unpredictable.
+#             kwargs (dict): Additional keyword arguments.
+#         """
+#         if pattern is None and map_func is None:
+#             raise ValueError("Invalid specification for Entity '%s'; no "
+#                              "pattern or mapping function provided. Either the"
+#                              " 'pattern' or the 'map_func' arguments must be "
+#                              "set." % name)
+#         self.name = name
+#         self.pattern = pattern
+#         self.mandatory = mandatory
+#         self.directory = directory
+#         self.map_func = map_func
+#         self.kwargs = kwargs
 
-        if (isinstance(dtype, six.string_types) and
-            dtype in ('str', 'float', 'int', 'bool')):
-            dtype = eval(dtype)
-        if dtype not in [str, float, int, bool, None]:
-            raise ValueError("Invalid dtype '%s'. Must be one of int, float, "
-                             "bool, or str." % dtype)
[…]
