### 2019-03-29 ee925ae
kill old classes and restructure modules
