+++ bids/layout/core.py
@@ -1,552 +1,0 @@
-""" Core pybids data structures. """
-
-import os
-import re
-import json
-from keyword import iskeyword
-import warnings
-from copy import deepcopy
-
-from .writing import build_path, write_contents_to_file
-from .models import BIDSFile, Entity, Tag
-from ..utils import listify, check_path_matches_patterns
-from ..config import get_option
-from ..external import six
-
-
-__all__ = [
-    "Config",
-    "Entity",
-    "BIDSFile",
-    "BIDSRootNode",
-    "BIDSSubjectNode",
-    "BIDSSessionNode"
-]
-
-
-class Config(object):
-    """ Container for BIDS configuration information.
-
-    Args:
-        name (str): The name to give the Config (e.g., 'bids').
-        entities (list): A list of dictionaries containing entity configuration
-            information.
-        default_path_patterns (list): Optional list of patterns used to build
-            new paths.
-    """
-
-    def __init__(self, name, entities=None, default_path_patterns=None):
-
-        self.name = name
-        self.entities = {}
-        self.default_path_patterns = default_path_patterns
-
-        if entities:
-            for ent in entities:
-                self.entities[ent['name']] = Entity(**ent)
-
-    @classmethod
-    def load(self, config):
-        if isinstance(config, six.string_types):
-            config_paths = get_option('config_paths')
-            if config in config_paths:
-                config = config_paths[config]
-            if not os.path.exists(config):
-                raise ValueError("{} is not a valid path.".format(config))
-            else:
-                with open(config, 'r') as f:
-                    config = json.load(f)
-        return Config(**config)
-
-
-# class Entity(object):
-
-#     def __init__(self, name, pattern=None, mandatory=False, directory=None,
-#                  map_func=None, dtype=None, **kwargs):
-#         """
-#         Represents a single entity defined in the JSON config.
-
-#         Args:
-#             name (str): The name of the entity (e.g., 'subject', 'run', etc.)
-#             pattern (str): A regex pattern used to match against file names.
-#                 Must define at least one group, and only the first group is
-#                 kept as the match.
-#             mandatory (bool): If True, every File _must_ match this entity.
-#             directory (str): Optional pattern defining a directory associated
-#                 with the entity.
-#             map_func (callable): Optional callable used to extract the Entity's
-#                 value from the passed string (instead of trying to match on the
-#                 defined .pattern).
-#             dtype (str): The optional data type of the Entity values. Must be
-#                 one of 'int', 'float', 'bool', or 'str'. If None, no type
-#                 enforcement will be attempted, which means the dtype of the
-#                 value may be unpredictable.
-#             kwargs (dict): Additional keyword arguments.
-#         """
-#         if pattern is None and map_func is None:
-#             raise ValueError("Invalid specification for Entity '%s'; no "
-#                              "pattern or mapping function provided. Either the"
-#                              " 'pattern' or the 'map_func' arguments must be "
-#                              "set." % name)
-#         self.name = name
-#         self.pattern = pattern
-#         self.mandatory = mandatory
-#         self.directory = directory
-#         self.map_func = map_func
-#         self.kwargs = kwargs
-
-#         if (isinstance(dtype, six.string_types) and
-#             dtype in ('str', 'float', 'int', 'bool')):
-#             dtype = eval(dtype)
-#         if dtype not in [str, float, int, bool, None]:
-#             raise ValueError("Invalid dtype '%s'. Must be one of int, float, "
-#                              "bool, or str." % dtype)
-#         self.dtype = dtype
-
-#         self.files = {}
-#         self.regex = re.compile(pattern) if pattern is not None else None
-
-#     def __iter__(self):
[…]
