### 2019-03-29 5b1273a
make sure dtypes are represented as strings before serialization
