### 2019-03-29 3f7860c
update get_file to work with DB models
