### 2019-04-05 #1294 @cMadan
@tyarkoni, how have things gone with getting people to confirm co-authorship? (E.g., how many have confirmed, how many have not responded)
