### 2019-04-05 #1294 @tyarkoni
Thanks for the reminder! I think ~3 people haven't responded (am on phone and can't easily look). I guess despite my earlier reservations, it may be best to just go ahead without those folks. Maybe I'll try to track them down via email if I can find one, but assuming I don't hear back in the next day or two, should we stick to the standard opt-in approach?
