### 2019-04-07 #1294 @cMadan
@tyarkoni, yes, that sounds like even more reason to stick with the standard opt-in approach. It would be  good to try and email those last few people in case they're not so active on GitHub. Let me know how it goes!
