### 2019-04-08 c560bba
add JSON handling to Tag; fix dtype conversion
