### 2019-04-08 3805e8f
~working reimplementation of get() method
