+++ bids/layout/layout.py
@@ -11,7 +11,7 @@ from bids_validator import BIDSValidator
 from ..utils import listify, natural_sort, check_path_matches_patterns
 from ..external import inflect, six
 from .writing import build_path, write_contents_to_file
-from .models import Base, Config, BIDSFile
+from .models import Base, Config, BIDSFile, Entity, Tag
 from .index import BIDSRootNode
 from .. import config as cf
 import sqlalchemy as sa
@@ -141,7 +141,8 @@ class BIDSLayout(object):
             documentation for the ignore argument for input format details.
             Note that paths in force_index takes precedence over those in
             ignore (i.e., if a file matches both ignore and force_index, it
-            *will* be indexed).
+            *will* be indexed). Note: NEVER include 'derivatives' here; use
+            the derivatives argument (or add_derivatives()) for that.
         config_filename (str): Optional name of filename within directories
             that contains configuration information.
         regex_search (bool): Whether to require exact matching (True) or regex
@@ -193,7 +194,8 @@ class BIDSLayout(object):
             config = 'bids'
         config = [Config.load(c) for c in listify(config)]
         self.config = {c.name: c for c in config}
-        self.root_node = BIDSRootNode(self.root, config, self)
+        self.root_node = BIDSRootNode(self.root, config, self,
+                                      index_metadata=True)
 
         # Consolidate entities into master list. Note: no conflicts occur b/c
         # multiple entries with the same name all point to the same instance.
@@ -451,10 +453,9 @@ class BIDSLayout(object):
         data.insert(0, 'path', [f.path for f in files])
         return data
 
-    def get(self, return_type='object', target=None, extensions=None,
-            scope='all', regex_search=False, defined_fields=None,
-            absolute_paths=None,
-            **kwargs):
+    def get(self, return_type='object', target=None, scope='all',
+            regex_search=False, absolute_paths=None, drop_invalid_filters=True,
+            **filters):
         """
         Retrieve files and/or metadata from the current Layout.
 
@@ -467,8 +468,6 @@ class BIDSLayout(object):
                     a valid target.
             target (str): Optional name of the target entity to get results for
                 (only used if return_type is 'dir' or 'id').
-            extensions (str, list): One or more file extensions to filter on.
-                BIDSFiles with any other extensions will be excluded.
             scope (str, list): Scope of the search space. If passed, only
                 nodes/directories that match the specified scope will be
                 searched. Possible values include:
@@ -479,13 +478,13 @@ class BIDSLayout(object):
             regex_search (bool or None): Whether to require exact matching
                 (False) or regex search (True) when comparing the query string
                 to each entity.
-            defined_fields (list): Optional list of names of metadata fields
-                that must be defined in JSON sidecars in order to consider the
-                file a match, but which don't need to match any particular
-                value.
             absolute_paths (bool): Optional the instance wide option to either
                 report absolute or relative (to the top of the dataset) paths.
-            kwargs (dict): Any optional key/values to filter the entities on.
+            drop_invalid_filters (bool): If False, any invalid filters (i.e.,
+                non-existent entities/keywords) are included (in which case the
+                query will always return nothing). If True, implicitly strips
+                invalid filters.
+            filters (dict): Any optional key/values to filter the entities on.
                 Keys are entity names, values are regexes to filter on. For
                 example, passing filter={'subject': 'sub-[12]'} would return
                 only files that match the first two subjects.
@@ -495,26 +494,17 @@ class BIDSLayout(object):
         """
 
         # Warn users still expecting 0.6 behavior
-        if 'type' in kwargs:
[…]
