### 2019-04-08 1e30148
fix failing test and temporarily comment out unimplemented test
