### 2019-04-09 6e9997a
comment out MetadataIndex and make sure session gets passed to Config.load()
