+++ bids/layout/layout.py
@@ -167,7 +167,7 @@ class BIDSLayout(object):
         self.derivatives = {}
         self.sources = sources
         self.regex_search = regex_search
-        self.metadata_index = MetadataIndex(self)
+        # self.metadata_index = MetadataIndex(self)
         self.config_filename = config_filename
         self.files = {}
         self.nodes = []
@@ -192,7 +192,8 @@ class BIDSLayout(object):
         # Set up configs
         if config is None:
             config = 'bids'
-        config = [Config.load(c) for c in listify(config)]
+        config = [Config.load(c, session=self.session)
+                  for c in listify(config)]
         self.config = {c.name: c for c in config}
         self.root_node = BIDSRootNode(self.root, config, self,
                                       index_metadata=True)
@@ -716,16 +717,14 @@ class BIDSLayout(object):
         path = os.path.abspath(path)
 
         # Make sure we have a valid suffix
-        suffix = kwargs.get('suffix')
-        if not suffix:
+        if not kwargs.get('suffix'):
             f = self.get_file(path)
             if 'suffix' not in f.entities:
                 raise ValueError(
                     "File '%s' does not have a valid suffix, most "
                     "likely because it is not a valid BIDS file." % path
                 )
-            suffix = f.entities['suffix']
-        kwargs['suffix'] = suffix
+            kwargs['suffix'] = f.entities['suffix']
 
         # Collect matches for all entities
         entities = {}
@@ -739,7 +738,9 @@ class BIDSLayout(object):
             for k in ignore_strict_entities:
                 entities.pop(k, None)
 
+        # Get candidate files
         results = self.get(return_type='object', **kwargs)
+
         # Make a dictionary of directories --> contained files
         folders = defaultdict(list)
         for f in results:
@@ -1014,116 +1015,116 @@ class BIDSLayout(object):
                                root=self.root)
 
 
-class MetadataIndex(object):
-    """A simple dict-based index for key/value pairs in JSON metadata.
+# class MetadataIndex(object):
+#     """A simple dict-based index for key/value pairs in JSON metadata.
 
-    Args:
-        layout (BIDSLayout): The BIDSLayout instance to index.
-    """
+#     Args:
+#         layout (BIDSLayout): The BIDSLayout instance to index.
+#     """
 
-    def __init__(self, layout):
-        self.layout = layout
-        self.key_index = {}
-        self.file_index = defaultdict(dict)
+#     def __init__(self, layout):
+#         self.layout = layout
+#         self.key_index = {}
+#         self.file_index = defaultdict(dict)
 
-    def index_file(self, f, overwrite=False):
-        """Index metadata for the specified file.
+#     def index_file(self, f, overwrite=False):
+#         """Index metadata for the specified file.
 
-        Args:
-            f (BIDSFile, str): A BIDSFile or path to an indexed file.
-            overwrite (bool): If True, forces reindexing of the file even if
-                an entry already exists.
-        """
-        if isinstance(f, six.string_types):
-            f = self.layout.get_file(f)
+#         Args:
+#             f (BIDSFile, str): A BIDSFile or path to an indexed file.
+#             overwrite (bool): If True, forces reindexing of the file even if
+#                 an entry already exists.
+#         """
+#         if isinstance(f, six.string_types):
+#             f = self.layout.get_file(f)
 
-        if f.path in self.file_index and not overwrite:
-            return
+#         if f.path in self.file_index and not overwrite:
+#             return
 
-        if 'suffix' not in f.entities:  # Skip files without suffixes
-            return
+#         if 'suffix' not in f.entities:  # Skip files without suffixes
+#             return
 
-        md = self._get_metadata(f.path)
[…]
