### 2019-04-09 1135d5c
update get_metadata to query the DB
