### 2019-04-09 c91c0bf
track whether BIDSFiles are directories
