### 2019-04-09 259c27a
return existing Config from .load() if one exists
