### 2019-04-09 38a4cbd
mostly working purely DB-based implementation
