+++ bids/layout/layout.py
@@ -12,7 +12,7 @@ from ..utils import listify, natural_sort, check_path_matches_patterns
 from ..external import inflect, six
 from .writing import build_path, write_contents_to_file
 from .models import Base, Config, BIDSFile, Entity, Tag
-from .index import BIDSRootNode
+from .index import index_layout
 from .. import config as cf
 import sqlalchemy as sa
 
@@ -195,8 +195,8 @@ class BIDSLayout(object):
         config = [Config.load(c, session=self.session)
                   for c in listify(config)]
         self.config = {c.name: c for c in config}
-        self.root_node = BIDSRootNode(self.root, config, self,
-                                      index_metadata=True)
+
+        index_layout(self, config, None, index_metadata=True)
 
         # Consolidate entities into master list. Note: no conflicts occur b/c
         # multiple entries with the same name all point to the same instance.
@@ -601,23 +601,27 @@ class BIDSLayout(object):
             query = query.join(BIDSFile.tags)
             regex = kwargs.get('regex_search', False)
             for name, val in filters.items():
-                query = query.filter()
                 if regex:
                     query = (query.filter(BIDSFile.tags.any(entity_name=name))
                              .filter(BIDSFile.tags._value.op('REGEXP')(val)))
                 else:
-                    query = query.filter(BIDSFile.tags.any(entity_name=name,
-                                                           _value=val))
+                    if isinstance(val, (list, tuple)):
+                        subq = sa.and_(Tag.entity_name==name,
+                                       Tag._value.in_(val))
+                        query = query.filter(BIDSFile.tags.any(subq))
+                    else:
+                        query = query.filter(
+                            BIDSFile.tags.any(entity_name=name, _value=val))
 
-        # Apply scoping
-        scope = kwargs.get('scope')
-        if scope != 'all':
-            if scope == 'derivatives':
-                query = query.filter_by(derivatives=True)
-            else:
-                if scope == 'raw':
-                    scope = 'bids'
-                query = query.filter_by(scope=scope)
+        # # Apply scoping
+        # scope = kwargs.get('scope')
+        # if scope != 'all':
+        #     if scope == 'derivatives':
+        #         query = query.filter_by(derivatives=True)
+        #     else:
+        #         if scope == 'raw':
+        #             scope = 'bids'
+        #         query = query.filter_by(scope=scope)
 
         return query
 
@@ -1009,118 +1013,3 @@ class BIDSLayout(object):
         write_contents_to_file(path, contents=contents, link_to=link_to,
                                content_mode=content_mode, conflicts=conflicts,
                                root=self.root)
-
-
-# class MetadataIndex(object):
-#     """A simple dict-based index for key/value pairs in JSON metadata.
-
-#     Args:
-#         layout (BIDSLayout): The BIDSLayout instance to index.
-#     """
-
-#     def __init__(self, layout):
-#         self.layout = layout
-#         self.key_index = {}
-#         self.file_index = defaultdict(dict)
-
-#     def index_file(self, f, overwrite=False):
-#         """Index metadata for the specified file.
-
-#         Args:
-#             f (BIDSFile, str): A BIDSFile or path to an indexed file.
-#             overwrite (bool): If True, forces reindexing of the file even if
-#                 an entry already exists.
-#         """
-#         if isinstance(f, six.string_types):
-#             f = self.layout.get_file(f)
-
-#         if f.path in self.file_index and not overwrite:
-#             return
-
-#         if 'suffix' not in f.entities:  # Skip files without suffixes
-#             return
-
-#         md = self._get_metadata(f.path)
-
-#         for md_key, md_val in md.items():
-#             if md_key not in self.key_index:
-#                 self.key_index[md_key] = {}
-#             self.key_index[md_key][f.path] = md_val
-#             self.file_index[f.path][md_key] = md_val
[…]
