+++ bids/layout/index.py
@@ -6,6 +6,7 @@ import json
 from keyword import iskeyword
 import warnings
 from copy import deepcopy
+from collections import defaultdict
 
 from .writing import build_path, write_contents_to_file
 from .models import Config, BIDSFile, Entity, Tag
@@ -14,197 +15,167 @@ from ..config import get_option
 from ..external import six
 
 
-class BIDSNode(object):
-    """ Represents a single directory or other logical grouping within a
-    BIDS project.
+def _extract_entities(bidsfile, entities):
+    match_vals = {}
+    for e in entities.values():
+        m = e.match_file(bidsfile)
+        if m is None and e.mandatory:
+            break
+        if m is not None:
+            match_vals[e.name] = (e, m)
+    return match_vals
 
-    Args:
-        path (str): The full path to the directory.
-        config (str, list): One or more names of built-in configurations
-            (e.g., 'bids' or 'derivatives') that specify the rules that apply
-            to this node.
-        root (BIDSNode): The node at the root of the tree the current node is
-            part of.
-        parent (BIDSNode): The parent of the current node.
-        force_index (bool): Whether or not to forcibly index every file below
-            this node, even if it fails standard BIDS validation.
-    """
 
-    _child_class = None
-    _child_entity = None
-    _entities = {}
+def index_layout(layout, config, force_index=None, index_metadata=True):
 
-    def __init__(self, path, config, entities=None, root=None, parent=None,
-                 force_index=False, index_metadata=True):
-        self.path = path
-        self.config = listify(config)
-        self.root = root
-        self.parent = parent
-        self.entities = {}
-        self.available_entities = {}
-        self.children = []
-        self.files = []
-        self.variables = []
-        self.force_index = force_index
+    session = layout.session
+    root_path = layout.root
+
+    # Track ALL entities we've seen in file names or metadata, starting with
+    # those found in the configs
+    all_entities = {}
+    for c in config:
+        all_entities.update(c.entities)
+
+    # For metadata indexing, we build up a store of all JSON data as we walk
+    # the file tree. It looks like: { dirname: { suffix: (entities, payload)}}
+    json_data = {}
+
+    def _index_file(f, dirpath, entities):
+        ''' Create DB record for file and its tags. '''
+        abs_fn = os.path.join(dirpath, f)
+
+        # Skip files that fail validation, unless forcibly indexing
+        if not force_index and not layout._validate_file(abs_fn):
+            return None, entities
+
+        bf = BIDSFile(abs_fn)
+        session.add(bf)
+
+        # Extract entity values
+        match_vals = {}
+        for e in entities.values():
+            m = e.match_file(bf)
+            if m is None and e.mandatory:
+                break
+            if m is not None:
+                match_vals[e.name] = (e, m)
+
+        # Create Entity <=> BIDSFile mappings
+        if match_vals:
+            for _, (ent, val) in match_vals.items():
+                tag = Tag(bf, ent, str(val), ent._dtype)
+                session.add(tag)
+
+        session.commit()
+
+        return bf, match_vals
+
+    def index_dir(path, config, parent=None, force_index=False):
+
+        abs_path = os.path.join(root_path, path)
+        config = list(config)     # Shallow copy
 
         # Check for additional config file in directory
-        layout_file = self.layout.config_filename
-        config_file = os.path.join(self.abs_path, layout_file)
+        layout_file = layout.config_filename
+        config_file = os.path.join(abs_path, layout_file)
         if os.path.exists(config_file):
-            cfg = Config.load(config_file, session=self.layout.session)
-            self.config.append(cfg)
+            cfg = Config.load(config_file, session=session)
+            config.append(cfg)
 
-        # Consolidate all entities
-        self._update_entities()
[…]
