### 2019-04-09 0df3f8d
improved efficiency in to_df() and added metadata flag
