### 2019-04-09 4c22cc1
remove Scope model in favor of multiple-DB solution
