### 2019-04-09 ce03f4b
add in_scope method to determine if layout is in passed scope
