### 2019-04-09 e79c5d5
add test of _in_scope
