### 2019-04-10 97fc7fe
fix _get_layouts_in_scope()
