### 2019-04-10 b8c12a1
reorder attributes and populate .entities from database
