### 2019-04-10 263a08f
add getters for files and entities
