### 2019-04-10 d8f8629
.entities and .files should have maximal scope by default
