### 2019-04-10 b5d6183
collect results from all in-scope layouts
