### 2019-04-10 836880d
allow leading periods in extension filter for backwards compatibility
