### 2019-04-10 a38169a
fix get_metadata--return values, not file paths
