### 2019-04-10 d5a707f
fix metadata indexing
