### 2019-04-10 80af17b
disallow list values when regex search is enabled
