### 2019-04-10 eb0511d
add regexp function and handle list values
