### 2019-04-10 df0ef7b
fix get_file query
