### 2019-04-10 bd59381
don't test copying until we figure out what to do about it
