### 2019-04-10 a682ec5
BIDSFiles no longer have access to layout, so update __repr__
