### 2019-04-10 3268893
replace .metadata property with get_metadata(), and fix DB call
