### 2019-04-10 183add1
fix get_nearest by making 'extension' default for ignore_strict_entities
