### 2019-04-10 db35c2c
eager load Tag and Entity associations in get() otherwise we have session management issues
