### 2019-04-10 80a48e6
fix force indexing--make sure arg is passed to indexing
