### 2019-04-10 c3e3af5
fixes to test_core, even though it's going to disappear soon
