### 2019-04-10 7b4f2cc
fix path writing tests
