+++ bids/layout/writing.py
@@ -2,7 +2,7 @@
 Contains helper functions that involve writing operations.
 '''
 
-import logging
+import warnings
 import os
 import re
 import sys
@@ -142,11 +142,11 @@ def write_contents_to_file(path, contents=None, link_to=None,
             raise ValueError(msg.format(path))
         elif conflicts == 'skip':
             msg = 'A file at path {} already exists, skipping writing file.'
-            logging.warn(msg.format(path))
+            warnings.warn(msg.format(path))
             return
         elif conflicts == 'overwrite':
             if isdir(path):
-                logging.warn('New path is a directory, not going to '
+                warnings.warn('New path is a directory, not going to '
                              'overwrite it, skipping instead.')
                 return
             os.remove(path)
