### 2019-04-10 8446a4d
fix get_metadata to accept scope arg
