### 2019-04-10 6f19781
move layout fixtures into conftest.py
