### 2019-04-11 6b1faee
add get_df() method to BIDSFile
