### 2019-04-11 7ce8df1
defines association relation; see #431
