### 2019-04-11 44c5b7f
enable filtering by association type
