### 2019-04-11 476485c
refactor indexing--separate metadata indexing for readability
