### 2019-04-11 096ba0d
create FileAssociations for metadata
