### 2019-04-11 e2ecbf2
create FileAssociation records for inheritance
