+++ bids/layout/index.py
@@ -135,65 +135,88 @@ def _index_metadata(layout):
 
     session = layout.session
     all_files = layout.get()
-    json_files = [bf for bf in all_files if bf.path.endswith('.json')]
-    filenames = [bf for bf in all_files if not bf.path.endswith('.json')]
-
-    # Wwe build up a store of all JSON data as we iterate files. It looks like:
-    # { dirname: { suffix: (entities, payload)}}
-    json_data = {}
 
     # Track ALL entities we've seen in file names or metadatas
     all_entities = {}
     for c in layout.config.values():
         all_entities.update(c.entities)
 
-    for bf in json_files:
+    # We build up a store of all file data as we iterate files. It looks like:
+    # { extension/suffix: dirname: [(entities, payload)]}}. The payload
+    # is left empty for non-JSON files.
+    file_data = {}
+
+    for bf in all_files:
         file_ents = bf.entities.copy()
         suffix = file_ents.pop('suffix', None)
-        file_ents.pop('extension', None)
+        ext = file_ents.pop('extension', None)
 
-        if suffix is not None:
-            with open(bf.path, 'r') as handle:
-                payload = json.load(handle)
-                if payload:
-                    to_store = (file_ents, payload, bf.path)
-                    if bf.dirname not in json_data:
-                        json_data[bf.dirname] = defaultdict(list)
-                    json_data[bf.dirname][suffix].append(to_store)
+        if suffix is not None and ext is not None:
+            key = "{}/{}".format(ext, suffix)
+            if key not in file_data:
+                file_data[key] = defaultdict(list)
 
+            if ext == 'json':
+                with open(bf.path, 'r') as handle:
+                    payload = json.load(handle)
+            else:
+                payload = None
+
+            to_store = (file_ents, payload, bf.path)
+            file_data[key][bf.dirname].append(to_store)
+
+
+    filenames = [bf for bf in all_files if not bf.path.endswith('.json')]
     for bf in filenames:
         file_ents = bf.entities.copy()
         suffix = file_ents.pop('suffix', None)
+        ext = file_ents.pop('extension', None)
         file_ent_keys = set(file_ents.keys())
 
-        if suffix is None:
+        if suffix is None or ext is None:
             continue
 
         # Extract metadata associated with the file. The idea is
         # that we loop over parent directories, and if we find
-        # payloads in the json_data store (indexing by directory
+        # payloads in the file_data store (indexing by directory
         # and current file suffix), we check to see if the
         # candidate JS file's entities are entirely consumed by
         # the current file. If so, it's a valid candidate, and we
         # add the payload to the stack. Finally, we invert the
         # stack and merge the payloads in order.
-        payloads = []
-        target = bf.dirname
-        while True:
-            if target in json_data and suffix in json_data[target]:
-                for js_ents, js_md, js_path in json_data[target][suffix]:
-                    js_keys = set(js_ents.keys())
-                    if (js_keys - file_ent_keys):
-                        continue
-                    matches = [js_ents[name] == file_ents[name]
-                                for name in js_keys]
-                    if all(matches):
-                        payloads.append((js_md, js_path))
+        ext_key = "{}/{}".format(ext, suffix)
+        json_key = "json/{}".format(suffix)
+        dirname = bf.dirname
 
-            parent = os.path.dirname(target)
-            if parent == target:
+        payloads = []
+        ancestors = []
+
+        while True:
+            # Get JSON payloads
+            json_data = file_data.get(json_key, {}).get(dirname, [])
+            for js_ents, js_md, js_path in json_data:
+                js_keys = set(js_ents.keys())
+                if (js_keys - file_ent_keys):
+                    continue
[…]
