### 2019-04-12 6dd1f8b
fix layout test and various minor tweaks
