### 2019-04-12 fc9c5ed
fix relative path layouts: always index using absolute paths
