### 2019-04-12 057a9a0
add option to include ancestor BIDSFiles in association list
