### 2019-04-12 726364a
allow either 'filename' or 'file' as get() return_type arg
