### 2019-04-12 e8a7e97
add BIDSLayout arg to control metadata indexing
