+++ bids/variables/tests/test_io.py
@@ -89,7 +89,8 @@ def test_load_synthetic_dataset(synthetic):
     # Sessions
     sessions = index.get_nodes('session')
     assert len(sessions) == 5 * 2
-    assert set(sessions[0].variables.keys()) == {'acq_time'}
+    target_cols = {'acq_time', 'TaskName', 'RepetitionTime'}
+    assert set(sessions[0].variables.keys()) == target_cols
     data = sessions[0].variables['acq_time'].filter({'task': 'nback'})
     assert len(data.values) == 2
 
