+++ bids/variables/io.py
@@ -123,14 +123,14 @@ def _load_time_variables(layout, dataset=None, columns=None, scan_length=None,
     '''
 
     # Extract any non-keyword arguments
-    kwargs = selectors.copy()
+    selectors = selectors.copy()
 
     if dataset is None:
         dataset = NodeIndex()
 
     selectors['datatype'] = 'func'
     selectors['suffix'] = 'bold'
-    images = layout.get(return_type='object', extension='.nii.gz',
+    images = layout.get(return_type='object', extension='nii.gz',
                         scope=scope, **selectors)
 
     if not images:
@@ -171,53 +171,50 @@ def _load_time_variables(layout, dataset=None, columns=None, scan_length=None,
 
         # Process event files
         if events:
-            dfs = layout.get_nearest(img_f, extension='.tsv', suffix='events',
-                                      all_=True, full_search=True,
-                                      ignore_strict_entities=['suffix'])
-            if dfs is not None:
-                for _data in dfs:
-                    _data = pd.read_csv(_data, sep='\t')
-                    if 'amplitude' in _data.columns:
-                        if (_data['amplitude'].astype(int) == 1).all() and \
-                                'trial_type' in _data.columns:
-                            msg = ("Column 'amplitude' with constant value 1 "
-                                   "is unnecessary in event files; ignoring "
-                                   "it.")
-                            _data = _data.drop('amplitude', axis=1)
-                        else:
-                            msg = ("Column name 'amplitude' is reserved; "
-                                   "renaming it to 'amplitude_'.")
-                            _data = _data.rename(
-                                columns={'amplitude': 'amplitude_'})
-                        warnings.warn(msg)
+            dfs = layout.get_nearest(
+                img_f, extension='tsv', suffix='events', all_=True,
+                full_search=True, ignore_strict_entities=['suffix', 'extension'])
+            for _data in dfs:
+                _data = pd.read_csv(_data, sep='\t')
+                if 'amplitude' in _data.columns:
+                    if (_data['amplitude'].astype(int) == 1).all() and \
+                            'trial_type' in _data.columns:
+                        msg = ("Column 'amplitude' with constant value 1 "
+                                "is unnecessary in event files; ignoring it.")
+                        _data = _data.drop('amplitude', axis=1)
+                    else:
+                        msg = ("Column name 'amplitude' is reserved; "
+                                "renaming it to 'amplitude_'.")
+                        _data = _data.rename(
+                            columns={'amplitude': 'amplitude_'})
+                    warnings.warn(msg)
 
-                    _data = _data.replace('n/a', np.nan)  # Replace BIDS' n/a
-                    _data = _data.apply(pd.to_numeric, errors='ignore')
+                _data = _data.replace('n/a', np.nan)  # Replace BIDS' n/a
+                _data = _data.apply(pd.to_numeric, errors='ignore')
 
-                    _cols = columns or list(set(_data.columns.tolist()) -
-                                            {'onset', 'duration'})
+                _cols = columns or list(set(_data.columns.tolist()) -
+                                        {'onset', 'duration'})
 
-                    # Construct a DataFrame for each extra column
-                    for col in _cols:
-                        df = _data[['onset', 'duration']].copy()
-                        df['amplitude'] = _data[col].values
+                # Construct a DataFrame for each extra column
+                for col in _cols:
+                    df = _data[['onset', 'duration']].copy()
+                    df['amplitude'] = _data[col].values
 
-                        # Add in all of the run's entities as new columns for
-                        # index
-                        for entity, value in entities.items():
-                            if entity in ALL_ENTITIES:
-                                df[entity] = value
[…]
