### 2019-04-12 594655a
better handling of duplicate entity columns
