+++ bids/variables/io.py
@@ -356,10 +356,13 @@ def _load_tsv_variables(layout, suffix, dataset=None, columns=None,
             ent_recs = [dict(layout.files[p].entities) for p in paths
                         if p in layout.files]
             ent_cols = pd.DataFrame.from_records(ent_recs)
+
+            # Remove entity columns found in both DFs
+            dupes = list(set(ent_cols.columns) & set(_data.columns))
+            to_drop = ['extension'] + dupes
+            ent_cols.drop(columns=to_drop, inplace=True)
+
             _data = pd.concat([_data, ent_cols], axis=1, sort=True)
-            _data.drop(columns=['extension'], inplace=True)
-            # It's possible to end up with duplicate entity columns this way
-            _data = _data.T.drop_duplicates().T
 
         # The BIDS spec requires ID columns to be named 'session_id', 'run_id',
         # etc., and IDs begin with entity prefixes (e.g., 'sub-01'). To ensure
