+++ bids/layout/tests/test_core.py
@@ -1,170 +1,0 @@
-import pytest
-import os
-import copy
-
-from bids.layout.models import Config, BIDSFile, Entity
-from bids import BIDSLayout
-from bids.tests import get_test_data_path
-
-
-DIRNAME = os.path.dirname(__file__)
-
-
-@pytest.fixture
-def sample_bidsfile(tmpdir):
-    testfile = 'sub-03_ses-2_task-rest_acq-fullbrain_run-2_bold.nii.gz'
-    fn = tmpdir.mkdir("tmp").join(testfile)
-    fn.write('###')
-    return BIDSFile(os.path.join(str(fn)))
-
-
-@pytest.fixture(scope='module')
-def subject_entity():
-    return Entity('subject', "[/\\\\]sub-([a-zA-Z0-9]+)", False,
-               "{subject}")
-
-
-def test_config_init_bare():
-    config = Config('custom')
-    assert config.name == 'custom'
-    assert config.entities == {}
-    assert config.default_path_patterns is None
-
-
-def test_config_init_with_args():
-    ents = [
-        {
-            "name": "task",
-            "pattern": "[_/\\\\]task-([a-zA-Z0-9]+)"
-        },
-        {
-            "name": "acquisition",
-            "pattern": "[_/\\\\]acq-([a-zA-Z0-9]+)"
-        }
-    ]
-    patterns = ['this_will_never_match_anything', 'and_neither_will_this']
-    config = Config('custom', entities=ents, default_path_patterns=patterns)
-    assert config.name == 'custom'
-    assert {'task', 'acquisition'} == set(config.entities.keys())
-    assert config.default_path_patterns  == patterns
-
-
-def test_config_init_from_class_load_bids():
-    config = Config.load('bids')
-    assert config.name == 'bids'
-    ent_names = ('subject', 'run', 'suffix')
-    assert all([en in config.entities for en in ent_names])
-    assert 'space' not in config.entities
-    first_patt = 'sub-{subject}[/ses-{session}]/anat/sub-{subject}'
-    assert config.default_path_patterns[0].startswith(first_patt)
-
-
-def test_config_init_from_class_load_derivatives():
-    config = Config.load('derivatives')
-    assert config.name == 'derivatives'
-    ent_names = ('space', 'atlas', 'roi')
-    assert all([en in config.entities for en in ent_names])
-    assert 'subject' not in config.entities
-    assert config.default_path_patterns is None
-
-
-def test_entity_init_minimal():
-    e = Entity('avaricious', r'aardvark-(\d+)')
-    assert e.name == 'avaricious'
-    assert e.pattern == r'aardvark-(\d+)'
-    assert not e.mandatory
-    assert e.directory is None
-    assert e.files == {}
-
-
-def test_entity_init_all_args(subject_entity):
-    ent = subject_entity
-    assert ent.name == 'subject'
-    assert ent.pattern == "[/\\\\]sub-([a-zA-Z0-9]+)"
-    assert ent.mandatory == False
-    assert ent.directory == "{subject}"
-
-
-def test_entity_init_with_bad_dtype():
-    with pytest.raises(ValueError) as exc:
-        ent = Entity('test', dtype='superfloat')
-        msg = exc.value.message
-        assert msg.startswith("Invalid dtype")
-
-
-def test_entity_deepcopy(subject_entity):
-    e = subject_entity
-    clone = copy.deepcopy(subject_entity)
-    for attr in ['name', 'pattern', 'mandatory', 'directory', 'regex']:
-        assert getattr(e, attr) == getattr(clone, attr)
[…]
