+++ bids/layout/tests/test_models.py
@@ -132,7 +132,8 @@ def test_config_init_with_args():
     patterns = ['this_will_never_match_anything', 'and_neither_will_this']
     config = Config('custom', entities=ents, default_path_patterns=patterns)
     assert config.name == 'custom'
-    assert [ent.name for ent in config.entities.values()] == ['task', 'acquisition']
+    target = {'task', 'acquisition'}
+    assert set(ent.name for ent in config.entities.values()) == target
     assert config.default_path_patterns  == patterns
 
 
