+++ bids/layout/tests/test_writing.py
@@ -44,7 +44,10 @@ def tmp_bids(tmpdir_factory):
     yield tmp_bids
     shutil.rmtree(str(tmp_bids))
     # Ugly hack
-    shutil.rmtree(join(get_test_data_path(), '7t_trt', 'sub-Bob'))
+    try:
+        shutil.rmtree(join(get_test_data_path(), '7t_trt', 'sub-Bob'))
+    except:
+        pass
 
 
 @pytest.fixture(scope='module')
