### 2019-04-19 dd95b08
add arguments controlling location/reindexing of SQLite database
