### 2019-04-19 9141555
conditionalize reindexing
