### 2019-04-19 a0c321f
add custom REGEXP function on begin event
