### 2019-04-19 5cca84e
make sure all test fixtures re-index datasets
