+++ bids/layout/tests/conftest.py
@@ -10,53 +10,54 @@ from bids.tests import get_test_data_path
 @pytest.fixture(scope="module")
 def layout_7t_trt():
     data_dir = join(get_test_data_path(), '7t_trt')
-    return BIDSLayout(data_dir)
+    return BIDSLayout(data_dir, reset_database=True)
 
 
 @pytest.fixture(scope="module")
 def layout_7t_trt_relpath():
     data_dir = join(get_test_data_path(), '7t_trt')
-    return BIDSLayout(data_dir, absolute_paths=False)
+    return BIDSLayout(data_dir, absolute_paths=False, reset_database=True)
 
 
 @pytest.fixture(scope="module")
 def layout_ds005():
     data_dir = join(get_test_data_path(), 'ds005')
-    return BIDSLayout(data_dir)
+    return BIDSLayout(data_dir, reset_database=True)
 
 
 @pytest.fixture(scope="module")
 def layout_ds117():
     data_dir = join(get_test_data_path(), 'ds000117')
-    return BIDSLayout(data_dir)
+    return BIDSLayout(data_dir, reset_database=True)
 
 
 @pytest.fixture(scope="module")
 def layout_ds005_derivs():
     data_dir = join(get_test_data_path(), 'ds005')
-    layout = BIDSLayout(data_dir)
+    layout = BIDSLayout(data_dir, reset_database=True)
     deriv_dir = join(data_dir, 'derivatives', 'events')
-    layout.add_derivatives(deriv_dir)
+    layout.add_derivatives(deriv_dir, reset_database=True)
     return layout
 
 
 @pytest.fixture(scope="module")
 def layout_ds005_multi_derivs():
     data_dir = join(get_test_data_path(), 'ds005')
-    layout = BIDSLayout(data_dir)
+    layout = BIDSLayout(data_dir, reset_database=True)
     deriv_dir1 = join(get_test_data_path(), 'ds005_derivs')
     deriv_dir2 = join(data_dir, 'derivatives', 'events')
-    layout.add_derivatives([deriv_dir1, deriv_dir2])
+    layout.add_derivatives([deriv_dir1, deriv_dir2], reset_database=True)
     return layout
 
 
 @pytest.fixture(scope="module")
 def layout_ds005_models():
     data_dir = join(get_test_data_path(), 'ds005')
-    return BIDSLayout(data_dir, validate=True, force_index=['models'])
+    return BIDSLayout(data_dir, validate=True, force_index=['models'],
+                      reset_database=True)
 
 
 @pytest.fixture(scope="module")
 def layout_synthetic():
     path = join(get_test_data_path(), 'synthetic')
-    return BIDSLayout(path, derivatives=True)
+    return BIDSLayout(path, derivatives=True, reset_database=True)
