### 2019-04-19 85f9767
revert to using in-memory DB by default
