### 2019-04-19 5f96e3b
add ability to save in-memory DB to file via .save()
