### 2019-04-20 e7b2816
fixes BIDSVariable deepcopy error by ensuring we have no connection to the DB
