### 2019-04-20 604d560
add deprecation warning for 'extensions' arg to get and improve docstrings
