### 2019-04-20 30767c8
update models.py docstrings
