### 2019-04-21 1931bb2
reorganize indexing/validation code for better readability/modularity
