+++ bids/utils.py
@@ -87,14 +87,3 @@ def splitext(path):
     li.extend(extensions)
     return li
 
-
-def check_path_matches_patterns(path, patterns):
-    ''' Check if the path matches at least one of the provided patterns. '''
-    path = os.path.abspath(path)
-    for patt in patterns:
-        if isinstance(patt, six.string_types):
-            if path == patt:
-                return True
-        elif patt.search(path):
-            return True
-    return False
