+++ bids/layout/index.py
@@ -8,9 +8,11 @@ import warnings
 from copy import deepcopy
 from collections import defaultdict, namedtuple
 
+from bids_validator import BIDSValidator
+
 from .writing import build_path, write_contents_to_file
 from .models import Config, BIDSFile, Entity, Tag, FileAssociation
-from ..utils import listify, check_path_matches_patterns
+from ..utils import listify
 from ..config import get_option
 from ..external import six
 
@@ -26,22 +28,121 @@ def _extract_entities(bidsfile, entities):
     return match_vals
 
 
-def index_layout(layout, force_index=None, index_metadata=True):
+def _check_path_matches_patterns(path, patterns):
+    ''' Check if the path matches at least one of the provided patterns. '''
+    if not patterns:
+        return False
+    path = os.path.abspath(path)
+    for patt in patterns:
+        if isinstance(patt, six.string_types):
+            if path == patt:
+                patterns.remove(patt)
+                return True
+        elif patt.search(path):
+            return True
+    return False
 
-    session = layout.session
-    root_path = layout.root
-    config = list(layout.config.values())
 
-    def _index_file(f, dirpath, entities):
+class BIDSLayoutIndexer(object):
+
+    def __init__(self, layout):
+
+        self.layout = layout
+        self.session = layout.session
+        self.validate = layout.validate
+        self.root = layout.root
+        self.config_filename = layout.config_filename
+        self.validator = BIDSValidator(index_associated=True)
+
+        # Create copies of list attributes we'll modify during indexing
+        self.config = list(layout.config.values())
+        self.include_patterns = list(layout.force_index)
+        self.exclude_patterns = list(layout.ignore)
+
+    def _validate_dir(self, d):
+        if _check_path_matches_patterns(d, self.include_patterns):
+            return True
+        return not _check_path_matches_patterns(d, self.exclude_patterns)
+
+    def _validate_file(self, f):
+        # Validate a file.
+        if _check_path_matches_patterns(f, self.include_patterns):
+            return True
+
+        if _check_path_matches_patterns(f, self.exclude_patterns):
+            return False
+
+        if not self.validate:
+            return True
+
+        # Derivatives are currently not validated.
+        # TODO: raise warning the first time in a session this is encountered
+        if 'derivatives' in self.layout.config:
+            return True
+
+        # BIDS validator expects absolute paths, but really these are relative
+        # to the BIDS project root.
+        to_check = os.path.relpath(f, self.root)
+        to_check = os.path.join(os.path.sep, to_check)
+
+        return self.validator.is_bids(to_check)
+
+    def _index_dir(self, path, config):
+
+        abs_path = os.path.join(self.root, path)
+
+        # Derivative directories must always be added separately
+        # and passed as their own root, so terminate if passed.
+        if (not self._validate_dir(abs_path) or
+            abs_path.startswith(os.path.join(self.root, 'derivatives'))):
+            return
+
+        config = list(config)     # Shallow copy
+
+        # Check for additional config file in directory
+        layout_file = self.config_filename
+        config_file = os.path.join(abs_path, layout_file)
+        if os.path.exists(config_file):
+            cfg = Config.load(config_file, session=self.session)
+            config.append(cfg)
+
+        # Track which entities are valid in filenames for this directory
+        config_entities = {}
+        for c in config:
+            config_entities.update(c.entities)
+
+        for (dirpath, dirnames, filenames) in os.walk(path):
+
+            # If layout configuration file exists, delete it
+            if self.config_filename in filenames:
+                filenames.remove(self.config_filename)
+
+            for f in filenames:
+
+                bf = self._index_file(f, dirpath, config_entities)
+                if bf is None:
+                    continue
+
+            self.session.commit()
+
+            # Recursively index subdirectories
[…]
