+++ bids/layout/layout.py
@@ -7,17 +7,17 @@ from functools import reduce, partial
 from itertools import chain
 import copy
 import warnings
+import sqlite3
 
-from bids_validator import BIDSValidator
-from ..utils import listify, natural_sort, check_path_matches_patterns
+import sqlalchemy as sa
+from sqlalchemy.orm import joinedload
+
+from ..utils import listify, natural_sort
 from ..external import inflect, six
 from .writing import build_path, write_contents_to_file
 from .models import Base, Config, BIDSFile, Entity, Tag
-from .index import index_layout
+from .index import BIDSLayoutIndexer
 from .. import config as cf
-import sqlalchemy as sa
-from sqlalchemy.orm import joinedload
-import sqlite3
 
 try:
     from os.path import commonpath
@@ -112,8 +112,6 @@ class BIDSLayout(object):
             files defined in the "core" BIDS spec, as setting validate=True
             will lead files in supplementary folders like derivatives/, code/,
             etc. to be ignored.
-        index_associated (bool): Argument passed onto the BIDSValidator;
-            ignored if validate = False.
         absolute_paths (bool): If True, queries always return absolute paths.
             If False, queries return relative paths (for files and directories).
         derivatives (bool, str, list): Specifies whether and/or which
@@ -166,21 +164,18 @@ class BIDSLayout(object):
     _default_ignore = {"code", "stimuli", "sourcedata", "models",
                        "derivatives", re.compile(r'^\.')}
 
-    def __init__(self, root, validate=True, index_associated=True,
-                 absolute_paths=True, derivatives=False, config=None,
-                 sources=None, ignore=None, force_index=None,
-                 config_filename='layout_config.json', regex_search=False,
-                 database_file=None, reset_database=False,
+    def __init__(self, root, validate=True, absolute_paths=True,
+                 derivatives=False, config=None, sources=None, ignore=None,
+                 force_index=None, config_filename='layout_config.json',
+                 regex_search=False, database_file=None, reset_database=False,
                  index_metadata=True):
 
         self.root = root
-        self._validator = BIDSValidator(index_associated=index_associated)
         self.validate = validate
         self.absolute_paths = absolute_paths
         self.derivatives = {}
         self.sources = sources
         self.regex_search = regex_search
-        # self.metadata_index = MetadataIndex(self)
         self.config_filename = config_filename
         self.ignore = [os.path.abspath(os.path.join(self.root, patt))
                        if isinstance(patt, six.string_types) else patt
@@ -191,22 +186,29 @@ class BIDSLayout(object):
 
         self.database_file = database_file
         self.session = None
-        index_dataset = self._init_db(database_file, reset_database)
 
         # Do basic BIDS validation on root directory
         self._validate_root()
 
         # Initialize the BIDS validator and examine ignore/force_index args
-        self._setup_file_validator()
+        self._validate_force_index()
+
+        index_dataset = self._init_db(database_file, reset_database)
 
         if index_dataset:
-            # Create Config objects and index layout
+            # Create Config objects
             if config is None:
                 config = 'bids'
             config = [Config.load(c, session=self.session)
                     for c in listify(config)]
             self.config = {c.name: c for c in config}
-            index_layout(self, self.force_index, index_metadata=index_metadata)
+
+            # Index files and (optionally) metadata
+            indexer = BIDSLayoutIndexer(self)
+            indexer.index_files()
+            if index_metadata:
+                indexer.index_metadata()
+
         else:
             # Load Configs from DB
             self.config = {c.name: c for c in self.session.query(Config).all()}
@@ -216,12 +218,11 @@ class BIDSLayout(object):
             if derivatives is True:
                 derivatives = os.path.join(root, 'derivatives')
[…]
