### 2019-04-21 f0925b3
fix inclusion/exclusion handling and add tests
