### 2019-04-22 85cb8d3
refactor BIDSFile into hierarchy of classes
