### 2019-04-22 60af65b
fix BIDSDataFile test and clean up
