+++ paper/paper.md
@@ -32,33 +32,35 @@ authors:
     name: Jean-Baptiste Poline
     orcid: 0000-0002-9794-749X
   - affiliation: 7
+    name: Dmitry Petrov
+  - affiliation: 8
     name: Valérie Hayot-Sasson
     orcid: 0000-0002-4830-4535
-  - affiliation: 8
+  - affiliation: 9
     name: Dylan M Nielson
     orcid: 0000-0003-4613-6643
-  - affiliation: 9
+  - affiliation: 10
     name: Johan Carlin
     orcid: 0000-0003-0933-1239
-  - affiliation: 10
+  - affiliation: 11
     name: Gregory Kiar
     orcid: 0000-0001-8915-496X
-  - affiliation: 11
+  - affiliation: 12
     name: Kirstie Whitaker
     orcid: 0000-0001-8498-4059
-  - affiliation: 10
+  - affiliation: 11
     name: Elizabeth DuPre
     orcid: 0000-0003-1358-196X
-  - affiliation: 12
+  - affiliation: 13
     name: Adina Wagner
     orcid: 0000-0003-2917-3450
-  - affiliation: 13
+  - affiliation: 14
     name: Lee S Tirrell
     orcid: 0000-0001-9393-8361
-  - affiliation: 14
+  - affiliation: 15
     name: Mainak Jas
     orcid: 0000-0002-3199-9027
-  - affiliation: 12
+  - affiliation: 13
     name: Michael Hanke
     orcid: 0000-0001-6398-6370
   - affiliation: 2
@@ -67,28 +69,28 @@ authors:
   - affiliation: 2
     name: Oscar Esteban
     orcid: 0000-0001-8435-6191
-  - affiliation: 15
+  - affiliation: 16
     name: Stefan Appelhoff
     orcid: 0000-0001-8002-0877
-  - affiliation: 16
+  - affiliation: 17
     name: Chris Holdgraf
     orcid: 0000-0002-2391-0678
-  - affiliation: 17
+  - affiliation: 18
     name: Isla Staden
     orcid: 0000-0002-0795-1154
-  - affiliation: 18
+  - affiliation: 19
     name: Bertrand Thirion
     orcid: 0000-0001-5018-7895
-  - affiliation: 19
+  - affiliation: 20
     name: Dave F Kleinschmidt
     orcid: 0000-0002-7442-2762
-  - affiliation: 8
+  - affiliation: 9
     name: John A Lee
     orcid: 0000-0001-5884-4247
-  - affiliation: 16
+  - affiliation: 17
     name: Matteo Visconti di Oleggio Castello
     orcid: 0000-0001-7931-5272
-  - affiliation: 20
+  - affiliation: 21
     name: Michael P Notter
     orcid: 0000-0002-5866-047X
   - affiliation: 2
@@ -108,32 +110,34 @@ affiliations:
 - index: 6
   name: McGill University
 - index: 7
-  name: Concordia University
+  name: Institute for Problems of Information Transmission
 - index: 8
-  name: National Institute of Mental Health
+  name: Concordia University
 - index: 9
-  name: MRC Cognition and Brain Sciences Unit
+  name: National Institute of Mental Health
 - index: 10
-  name: Montreal Neurological Institute and Hospital
+  name: MRC Cognition and Brain Sciences Unit
 - index: 11
-  name: Alan Turing Institute
+  name: Montreal Neurological Institute and Hospital
 - index: 12
-  name: Otto-von-Guericke University Magdeburg
+  name: Alan Turing Institute
 - index: 13
-  name: CorticoMetrics LLC
+  name: Otto-von-Guericke University Magdeburg
 - index: 14
-  name: Télécom ParisTech, France
+  name: CorticoMetrics LLC
 - index: 15
[…]
