### 2019-04-30 eeb80a7
ugly-ass fix for Python 3.5 test failure, and cleanup
