### 2019-05-01 2af4650
include FileAssociation kind in uniqueness constraint
