### 2019-05-07 5e8308e
add get_entities method to BIDSFile and reorder attributes
