### 2019-05-07 2b68130
rename is_metadata argument to just metadata
