### 2019-05-07 a3b07d1
reorganize BIDSFile tests
