### 2019-05-07 7f689fe
add BIDSFile.get_entities() method and tests
