### 2019-05-10 29b1014
no need for JSON column when a String will do; closes #437
