### 2019-05-10 c8c6668
BF: initial fixups to bids/layout/README.md
also to automate testing it for at least correct execution with smth like

    grep '>>>' bids/layout/README.md | sed -e 's,>>> ,,g' >| /tmp/test-readme.py; python /tmp/test-readme.py
