### 2019-05-10 58dcac8
RF: "Kill" bids/layout/README.md and point to the tutorial
Not sure if linking would work out
