+++ bids/layout/README.md
@@ -1,152 +1,6 @@
 # layout
-Use grabbit to interact with BIDS projects
 
-## Overview
-
-[Grabbit](https://github.com/grabbles/grabbit) is a lightweight Python 2 and 3 package for simple queries over filenames within a project. It's geared towards projects or applications with highly structured filenames that allow useful queries to be performed without having to inspect the file metadata or contents. Layout is a BIDS-specific extension of grabbit that makes it particularly easy to work with BIDS projects, and provides additional functionality.
-
-## Quickstart
-
-Suppose we have a BIDS project directory that looks like this (partial listing of `./bids/tests/data/7t_trt` shipped within pybids):
-
-```
-├── dataset_description.json
-├── participants.tsv
-├── sub-01
-│   ├── ses-1
-│   │   ├── anat
-│   │   │   ├── sub-01_ses-1_T1map.nii.gz
-│   │   │   └── sub-01_ses-1_T1w.nii.gz
-│   │   ├── fmap
-│   │   │   ├── sub-01_ses-1_run-1_magnitude1.nii.gz
-│   │   │   ├── sub-01_ses-1_run-1_magnitude2.nii.gz
-│   │   │   ├── sub-01_ses-1_run-1_phasediff.json
-│   │   │   ├── sub-01_ses-1_run-1_phasediff.nii.gz
-│   │   │   ├── sub-01_ses-1_run-2_magnitude1.nii.gz
-│   │   │   ├── sub-01_ses-1_run-2_magnitude2.nii.gz
-│   │   │   ├── sub-01_ses-1_run-2_phasediff.json
-│   │   │   └── sub-01_ses-1_run-2_phasediff.nii.gz
-│   │   ├── func
-│   │   │   ├── sub-01_ses-1_task-rest_acq-fullbrain_run-1_bold.nii.gz
-│   │   │   ├── sub-01_ses-1_task-rest_acq-fullbrain_run-1_physio.tsv.gz
-│   │   │   ├── sub-01_ses-1_task-rest_acq-fullbrain_run-2_bold.nii.gz
-│   │   │   ├── sub-01_ses-1_task-rest_acq-fullbrain_run-2_physio.tsv.gz
-│   │   │   ├── sub-01_ses-1_task-rest_acq-prefrontal_bold.nii.gz
-│   │   │   └── sub-01_ses-1_task-rest_acq-prefrontal_physio.tsv.gz
-│   │   └── sub-01_ses-1_scans.tsv
-│   ├── ses-2
-│   │   ├── fmap
-│   │   │   ├── sub-01_ses-2_run-1_magnitude1.nii.gz
-│   │   │   ├── sub-01_ses-2_run-1_magnitude2.nii.gz
-│   │   │   ├── sub-01_ses-2_run-1_phasediff.json
-│   │   │   ├── sub-01_ses-2_run-1_phasediff.nii.gz
-│   │   │   ├── sub-01_ses-2_run-2_magnitude1.nii.gz
-│   │   │   ├── sub-01_ses-2_run-2_magnitude2.nii.gz
-│   │   │   ├── sub-01_ses-2_run-2_phasediff.json
-│   │   │   └── sub-01_ses-2_run-2_phasediff.nii.gz
-```
-
-We can initialize a `layout` Layout object like so:
-
-```python
->>> from bids.layout import BIDSLayout
->>> project_root = 'bids/tests/data/7t_trt'
->>> layout = BIDSLayout(project_root)
-```
-
[…]
