### 2019-05-10 #1294 @tyarkoni
@cMadan apologies for the delay: the author list is now finalized. I think we're good to go. We have a fairly major PyBIDS release coming up (0.9), so we could in theory wait on that, but I don't know that it matters much (it won't change the text in any way).
