### 2019-05-10 b518712
update CHANGELOG in prep for release
