+++ bids/layout/README.md
@@ -1,6 +1,6 @@
 # layout
 
 PyBIDS used to depend on [Grabbit](https://github.com/grabbles/grabbit) and here
-we described interactions with its Layout.  Since them PyBIDS grew into an
+we described interactions with its Layout.  Since then PyBIDS grew into an
 independent beast and we refer the curious minds to
 [our tutorial](../../examples/pybids_tutorial.ipynb) Jupyter notebook.
