### 2019-05-11 #1294 @cMadan
Thanks for getting back to this, Tal. In that case I think it would be better to make 0.9 the version associated with the JOSS paper.

I also looked over the paper and it looks like the bib needs a bit of editing:
- Fix Gorgolweski's initials across the entries so they are formatted the same. Currently this is not the case, which is why the PDF prints the first initials in the citations.
- Spell out full journal names in all references and capitalise appropriately.
- Capitalise "MRI" in article titles.
- Since we will have the Zenodo archive DOI on the front page of the paper, as part of the metadata shown in the margin note, we don't include a citation to it or mention it in the paper. Can you delete that reference?

Also:
- Add space after close parentheses in "fitlins)--", otherwise URL gets broken.
