### 2019-05-14 c9f02d4
DOC: Returns -> Return to stay inline with imperative descriptions of other functions
