### 2019-05-14 95a1fef
DOC: minor rewording of compute_regressor docstring 1 line summary
