### 2019-05-14 f41d701
RF: strip away unused derivatives=False, is_dir=False from BIDSFile signature
Seems to be not stored/used -- may be was left for compatibility?  Then
should gain some comment
