### 2019-05-14 f5602a8
DOC: Docstring summaries for BIDSFile derived classes
