### 2019-05-14 7cc2929
adds BIDSJSONFile; closes #442
