### 2019-05-14 d133f1c
ENH: convert relative path to absolute relative to the top of layout in get_metadata
