### 2019-05-15 f73379d
add 'self' option to _get_layouts_in_scope
