### 2019-05-15 53fb7ba
add get_dataset_description method
