+++ bids/layout/tests/test_layout.py
@@ -495,6 +495,18 @@ def test_get_layouts_in_scope(layout_ds005_multi_derivs):
     assert self_scope == [l]
 
 
+def test_get_dataset_description(layout_ds005_multi_derivs):
+    l = layout_ds005_multi_derivs
+    dd = l.get_dataset_description()
+    assert isinstance(dd, dict)
+    assert dd['Name'] == 'Mixed-gambles task'
+    dd = l.get_dataset_description('all', True)
+    assert isinstance(dd, list)
+    assert len(dd) == 3
+    names = {'Mixed-gambles task', 'Mixed-gambles task -- dummy derivative'}
+    assert set([d['Name'] for d in dd]) == names
+
+
 def test_indexed_file_associations(layout_7t_trt):
     img = layout_7t_trt.get(subject='01', run=1, suffix='bold', session='1',
                             acquisition='fullbrain', extension='nii.gz')[0]
