### 2019-05-22 daa48d7
CI: Check packages
