### 2019-05-23 555f6ed
MAINT: Move configuration to setup.py
