+++ long_description.rst
@@ -1,0 +1,16 @@
+PyBIDS
+======
+PyBIDS is a Python module to interface with datasets conforming BIDS.
+See BIDS paper_ and http://bids.neuroimaging.io website for more information.
+
+.. _paper: http://www.nature.com/articles/sdata201644
+
+License
+=======
+``pybids`` is licensed under the terms of the MIT license. See the file
+"LICENSE" for information on the history of this software, terms & conditions
+for usage, and a DISCLAIMER OF ALL WARRANTIES.
+
+All trademarks referenced herein are property of their respective holders.
+
+Copyright (c) 2016--, PyBIDS developers, Planet Earth
