+++ setup.cfg
@@ -1,5 +1,57 @@
 [metadata]
-description-file = README.md
+name = pybids
+url = "http://github.com/bids-standard/pybids"
+author = PyBIDS Developers
+author_email = bids-discussion@googlegroups.com
+maintainer = PyBIDS Developers
+maintainer_email = bids-discussion@googlegroups.com
+description = bids: interface with datasets conforming to BIDS
+long_description = file:long_description.rst
+long_description_content_type = text/x-rst; charset=UTF-8
+license = MIT
+platforms = OS Independent
+classifiers =
+    Development Status :: 3 - Alpha
+    Environment :: Console
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Topic :: Scientific/Engineering
+
+[options]
+python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*
+install_requires =
+    numpy
+    scipy
+    nibabel >=2.1
+    pandas >=0.23
+    patsy
+    sqlalchemy
+    bids-validator
+    num2words
+tests_require =
+    pytest >=3.3
+    mock
+    pathlib ; python_version < "3.4"
+packages = find:
+include_package_data = True
+
+[options.extras_require]
+analysis =
+test =
+    pytest >=3.3
+    mock
+    pathlib ; python_version < "3.4"
+travis_tests =
+    %(test)s
+    codecov
+    pytest-cov
+    pytest-xdist
+tutorial =
+    nbconvert
+    jupyter_client
+    ipykernel
 
 [versioneer]
 VCS = git
