### 2019-05-23 20b3d86
MNT: Quiet flag for git clone
