### 2019-05-24 99bfd50
TEST: Verify Path can accept a BIDSFile
