### 2019-05-24 fd7a4ef
DOC: Add PyPI badge for latest version available
