### 2019-06-26 2fc9beb
DOC: Update bibliography file with eye to formatting
