+++ paper/paper.bib
@@ -43,7 +43,8 @@
 @article{Gorgolewski2016-sk,
   title    = "A Practical Guide for Improving Transparency and Reproducibility
               in Neuroimaging Research",
-  author   = "Gorgolewski, Krzysztof J and Poldrack, Russell A",
+  author   = {Gorgolewski, Krzysztof J and
+              Poldrack, Russell A},
   abstract = "Recent years have seen an increase in alarming signals regarding
               the lack of replicability in neuroscience, psychology, and other
               related fields. To avoid a widespread crisis in neuroimaging
@@ -57,7 +58,7 @@
               advice as well as highlighting advantages of adopting more open
               research practices that go beyond improved transparency and
               reproducibility.",
-  journal  = "PLoS Biol.",
+  journal  = "PLOS Biology",
   volume   =  14,
   number   =  7,
   pages    = "e1002506",
@@ -70,15 +71,33 @@
 @article{Gorgolewski2016-nw,
   title    = "The brain imaging data structure, a format for organizing and
               describing outputs of neuroimaging experiments",
-  author   = "Gorgolewski, Krzysztof J and Auer, Tibor and Calhoun, Vince D and
-              Craddock, R Cameron and Das, Samir and Duff, Eugene P and
-              Flandin, Guillaume and Ghosh, Satrajit S and Glatard, Tristan and
-              Halchenko, Yaroslav O and Handwerker, Daniel A and Hanke, Michael
-              and Keator, David and Li, Xiangrui and Michael, Zachary and
-              Maumet, Camille and Nichols, B Nolan and Nichols, Thomas E and
-              Pellman, John and Poline, Jean-Baptiste and Rokem, Ariel and
-              Schaefer, Gunnar and Sochat, Vanessa and Triplett, William and
-              Turner, Jessica A and Varoquaux, Ga{\"e}l and Poldrack, Russell A",
+  author   = {Gorgolewski, Krzysztof J and
+              Auer, Tibor and
+              Calhoun, Vince D and
+              Craddock, R Cameron and
+              Das, Samir and
+              Duff, Eugene P and
+              Flandin, Guillaume and
+              Ghosh, Satrajit S and
+              Glatard, Tristan and
+              Halchenko, Yaroslav O and
+              Handwerker, Daniel A and
+              Hanke, Michael and
+              Keator, David and
+              Li, Xiangrui and
+              Michael, Zachary and
+              Maumet, Camille and
+              Nichols, B Nolan and
+              Nichols, Thomas E and
+              Pellman, John and
+              Poline, Jean-Baptiste and
+              Rokem, Ariel and
+              Schaefer, Gunnar and
+              Sochat, Vanessa and
+              Triplett, William and
+              Turner, Jessica A and
+              Varoquaux, Ga{\"e}l and
+              Poldrack, Russell A},
   abstract = "The development of magnetic resonance imaging (MRI) techniques
               has defined modern neuroimaging. Since its inception, tens of
               thousands of studies using techniques such as functional MRI and
@@ -96,7 +115,7 @@
               unifies the majority of practices already common in the field,
               and captures the metadata necessary for most common data
               processing operations.",
-  journal  = "Sci Data",
+  journal  = "Scientific Data",
   volume   =  3,
   pages    = "160044",
   month    =  jun,
@@ -108,17 +127,35 @@
 @article{Gorgolewski2017-cz,
   title    = "{BIDS} apps: Improving ease of use, accessibility, and
               reproducibility of neuroimaging data analysis methods",
[…]
