### 2019-06-26 d749439
DOC: Drop Zenodo reference
