+++ paper/paper.bib
@@ -1,45 +1,3 @@
-@misc{zenodo,
-  author       = {Yarkoni, Tal and
-                  Markiewicz, Christopher J. and
-                  de la Vega, Alejandro and
-                  Gorgolewski, Krzysztof J. and
-                  Salo, Taylor and
-                  Halchenko, Yaroslav O. and
-                  McNamara, Quinten and
-                  DeStasio, Krista and
-                  Poline, Jean-Baptiste and
-                  Petrov, Dmitry and
-                  Hayot-Sasson, Valérie and
-                  Nielson, Dylan M. and
-                  Carlin, Johan and
-                  Kiar, Gregory and
-                  Whitaker, Kirstie and
-                  DuPre, Elizabeth and
-                  Wagner, Adina and
-                  Ivanov, Alexander and
-                  Tirrell, Lee S. and
-                  Jas, Mainak and
-                  Hanke, Michael and
-                  Poldrack, Russell and
-                  Esteban, Oscar and
-                  Appelhoff, Stefan and
-                  Holdgraf, Chris and
-                  Staden, Isla and
-                  Rokem, Ariel and
-                  Thirion, Bertrand and
-                  Kleinschmidt, Dave F. and
-                  Lee, John A. and
-                  Visconti di Oleggio Castello, Matteo and
-                  Notter, Michael Philipp and
-                  Roca, Pauline and
-                  Blair, Ross},
-  title        = {bids-standard/pybids: 0.8.0},
-  month        = feb,
-  year         = 2019,
-  doi          = {10.5281/zenodo.2566511},
-  url          = {https://doi.org/10.5281/zenodo.2566511}
-}
-
 @article{Gorgolewski2016-sk,
   title    = "A Practical Guide for Improving Transparency and Reproducibility
               in Neuroimaging Research",
