### 2019-06-26 828bdb7
DOC: Spaces around emdashes, italicize e.g., i.e.
