### 2019-06-26 #1294 @effigies
@kyleniemeyer @cMadan I went ahead and addressed these concerns: https://github.com/bids-standard/pybids/pull/452

If possible, it would be nice to try to render before merging to master, to avoid too many PRs. I don't know if there are docs on how to render locally. The main thing I'm not sure of is whether using the bracket (`{}`) syntax will resolve the issues with KJG's name in citations.
