### 2019-06-26 #1294 @kyleniemeyer
@effigies you might check out the `compile` script another JOSS author put together, which I've used successfully to locally build a paper fairly easily: https://github.com/connorourke/crystal_torture/blob/master/paper/compile
