### 2019-06-26 #1294 @cMadan
@effigies, it's possible to generate the pdf from a branch `@whedon generate pdf from branch custom-branch-name`, but it looks like your current PR is on a branch of a fork, which is not supported.
