### 2019-06-26 8551700
DOC: Fix BIDS citation
