### 2019-06-26 #1294 @effigies
@cMadan Thanks for the tip. Made one more change, so I'll re-render and I think we'll be ready for a final review, pending @tyarkoni approval of the changes I've made.
