### 2019-06-26 #1294 @effigies
New PR: https://github.com/bids-standard/pybids/pull/453
