### 2019-06-27 2a89c7e
TEST: Expect failure on Python < 3.6
