### 2019-07-02 #1294 @cMadan
@tyarkoni, is this now ready for a final review?
