### 2019-07-05 7f33b5c
fix typo in docstring
