### 2019-07-05 ca12787
pep8 fixes
