+++ bids/layout/layout.py
@@ -1,9 +1,10 @@
+"""BIDSLayout class."""
 import os
 import json
 import re
 from collections import defaultdict
 from io import open
-from functools import reduce, partial
+from functools import partial
 from itertools import chain
 import copy
 import warnings
@@ -15,8 +16,7 @@ from sqlalchemy.orm import joinedload
 from ..utils import listify, natural_sort, make_bidsfile
 from ..external import inflect, six
 from .writing import build_path, write_contents_to_file
-from .models import (Base, Config, BIDSFile, Entity, Tag, BIDSDataFile,
-                     BIDSImageFile)
+from .models import (Base, Config, BIDSFile, Entity, Tag)
 from .index import BIDSLayoutIndexer
 from .. import config as cf
 
@@ -35,7 +35,7 @@ __all__ = ['BIDSLayout']
 
 def parse_file_entities(filename, entities=None, config=None,
                         include_unmatched=False):
-    """ Parse the passed filename for entity/value pairs.
+    """Parse the passed filename for entity/value pairs.
 
     Args:
         filename (str): The filename to parse for entity values
@@ -52,7 +52,6 @@ def parse_file_entities(filename, entities=None, config=None,
     Returns: A dict, where keys are Entity names and values are the
         values extracted from the filename.
     """
-
     # Load Configs if needed
     if entities is None:
 
@@ -80,7 +79,7 @@ def parse_file_entities(filename, entities=None, config=None,
 
 
 def add_config_paths(**kwargs):
-    """ Add to the pool of available configuration files for BIDSLayout.
+    """Add to the pool of available configuration files for BIDSLayout.
 
     Args:
         kwargs: dictionary specifying where to find additional config files.
@@ -90,7 +89,6 @@ def add_config_paths(**kwargs):
         > add_config_paths(my_config='/path/to/config')
         > layout = BIDSLayout('/path/to/bids', config=['bids', 'my_config'])
     """
-
     for k, path in kwargs.items():
         if not os.path.exists(path):
             raise ValueError(
@@ -103,7 +101,7 @@ def add_config_paths(**kwargs):
 
 
 class BIDSLayout(object):
-    """ Layout class representing an entire BIDS dataset.
+    """Layout class representing an entire BIDS dataset.
 
     Args:
         root (str): The root directory of the BIDS dataset.
@@ -114,7 +112,8 @@ class BIDSLayout(object):
             will lead files in supplementary folders like derivatives/, code/,
             etc. to be ignored.
         absolute_paths (bool): If True, queries always return absolute paths.
-            If False, queries return relative paths (for files and directories).
+            If False, queries return relative paths (for files and
+            directories).
         derivatives (bool, str, list): Specifies whether and/or which
             derivatives to to index. If True, all pipelines found in the
             derivatives/ subdirectory will be indexed. If a str or list, gives
@@ -170,7 +169,7 @@ class BIDSLayout(object):
                  force_index=None, config_filename='layout_config.json',
                  regex_search=False, database_file=None, reset_database=False,
                  index_metadata=True):
-
+        """Initialize BIDSLayout."""
         self.root = root
         self.validate = validate
         self.absolute_paths = absolute_paths
@@ -201,7 +200,7 @@ class BIDSLayout(object):
             if config is None:
                 config = 'bids'
             config = [Config.load(c, session=self.session)
-                    for c in listify(config)]
+                      for c in listify(config)]
             self.config = {c.name: c for c in config}
 
             # Index files and (optionally) metadata
@@ -226,8 +225,8 @@ class BIDSLayout(object):
                 index_metadata=index_metadata)
 
     def __getattr__(self, key):
-        ''' Dynamically inspect missing methods for get_<entity>() calls
[…]
