### 2019-07-09 2067056
FIX: Instantiate ignore/force_index after root validation
