### 2019-07-09 e293ae2
TEST: Verify entity=None fails to find files lacking that entity
