+++ bids/layout/tests/test_layout.py
@@ -222,6 +222,13 @@ def test_get_return_type_dir(layout_7t_trt, layout_7t_trt_relpath):
     assert target == res2
 
 
+def test_get_val_none(layout_7t_trt):
+    t1w_files = layout_7t_trt.get(subject='01', ses='1', suffix='T1w')
+    assert len(t1w_files) == 1
+    assert 'acq' not in t1w_files[0].path
+    t1w_files = layout_7t_trt.get(subject='01', ses='1', suffix='T1w', acquisition=None)
+    assert len(t1w_files) == 1
+
 
 def test_get_return_sorted(layout_7t_trt):
     bids_files = layout_7t_trt.get(target='subject')
