### 2019-07-10 ff30fa4
FIX: Allow None to select for the absence of a tag
