### 2019-07-10 d61af90
tst(BIDSLayout): add two more tests to the entity=None case
