### 2019-07-10 0c39d71
fix(BIDSLayout): more readable query building
