### 2019-07-11 00f4e08
MAINT: Update changelog, Zenodo ordering
