### 2019-07-12 090ec03
CI: Fix deployment conditions
