### 2019-07-12 2b27ed6
CI: Restrict deployment to tags
