### 2019-07-30 #1294 @cMadan
@tyarkoni @effigies, everything looks good to me!

To move forward with accepting your submission, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
