### 2019-07-30 #1294 @effigies
Hi @cMadan I'm not sure I understand. We have a release [0.9.2](https://github.com/bids-standard/pybids/releases/tag/0.9.2) and a Zenodo (https://doi.org/10.5281/zenodo.3333492). I think the author list is slightly different because not all could be contacted to give explicit consent to be on the JOSS paper. And the title is not the same.

I guess the idea is to create a separate release? Is there a naming convention to follow, e.g. `0.9.2-joss`? And I'm not sure what the threshold for too small of a contribution might be, but once identified, if we remove them from the Zenodo, presumably we need to remove them from the paper as well?
