### 2019-07-31 #1294 @cMadan
@effigies, I think those discrepencies are fine then.

Any thoughts, @openjournals/joss-eics ?
