### 2019-07-31 #1294 @labarba
JOSS requests that you make a new release of the software after applying any revisions that occurred during the review. That new release should be deposited in Zenodo, for archival of the reviewed software.

Our expectation is that the Zenodo archive corresponding to the accepted JOSS paper have the same title and author list as the JOSS paper.

If I understand correctly, you're saying that some of the contributors to the software are included in the author list of the Zenodo archive, but are not listed on the JOSS paper because you couldn't contact them to agree.

What is the difference between listing someone as author on a Zenodo archive and a JOSS paper? Is there any? Why can they be listed without their approval in the Zenodo archive, but not in the JOSS paper? I don't have answers to this. We do leave it to the authors to decide on their authorship. But it is a bit strange to have a different author list on an archive of a software and a paper about that same software.
