### 2019-07-31 #1294 @effigies
> JOSS requests that you make a new release of the software after applying any revisions that occurred during the review. That new release should be deposited in Zenodo, for archival of the reviewed software.

That's the case. All JOSS-related revisions are in the most recent release.

> Our expectation is that the Zenodo archive corresponding to the accepted JOSS paper have the same title and author list as the JOSS paper.

We can probably update the title... @tyarkoni do you have ownership of the Zenodo entry?

> What is the difference between listing someone as author on a Zenodo archive and a JOSS paper? Is there any? Why can they be listed without their approval in the Zenodo archive, but not in the JOSS paper?

We settled on opt-out for Zenodo in https://github.com/bids-standard/pybids/pull/308, although we did get a very high rate of affirmative response, with I think only 7 failures to respond. As to why we chose opt-out for Zenodo, the default behavior of Zenodo is to list all contributors in the git history (or possibly some other GitHub metadata...), so removing a contributor for failing to respond seemed punitive.

At an above request, we sought an additional explicit opt-in for JOSS in https://github.com/bids-standard/pybids/pull/395, with 3 failures to respond. I personally find the discrepancy acceptable. Software authorship credits do not have the same reputational implications of paper authorship, and JOSS seems to fall enough on the paper side of things to justify an active opt-in, which matches your policies.

----

However, I leave it to you. We can make a second tag on with only JOSS-consenting authors. If so, please let me know what a sensible tag would be.
