### 2019-07-31 #1294 @labarba
@arfon — Are we OK with an author-list discrepancy between the JOSS paper and Zenodo archive, for reasons described above?
