### 2019-07-31 #1294 @arfon
> @arfon — Are we OK with an author-list discrepancy between the JOSS paper and Zenodo archive, for reasons described above?

Yes, that’s fine.
