### 2019-08-04 ae1f268
make bids.reports deal with .nii only BIDS dataset
 - changes all extension='.nii.gz' to extention=[".nii", ".nii.gz"]
 - add missing import os in report/util.py
