+++ bids/reports/utils.py
@@ -5,6 +5,7 @@ methods section from a BIDS dataset.
 """
 from __future__ import print_function
 import logging
+import os
 from .. import __version__
 
 logging.basicConfig()
