+++ bids/reports/parsing.py
@@ -315,7 +315,7 @@ def fmap_info(metadata, img, config, layout):
         run_dict = {}
         for scan in scans:
             fn = basename(scan)
-            iff_file = [f for f in layout.get(extension='nii.gz') if fn in f.path][0]
+            iff_file = [f for f in layout.get(extension=[".nii", ".nii.gz"]) if fn in f.path][0]
             run_num = int(iff_file.run)
             ty = iff_file.entities['suffix'].upper()
             if ty == 'BOLD':
@@ -439,7 +439,7 @@ def parse_niftis(layout, niftis, subj, config, **kwargs):
 
             if nifti_struct.entities['datatype'] == 'func':
                 if not skip_task.get(nifti_struct.entities['task'], False):
-                    echos = layout.get_echoes(subject=subj, extension='nii.gz',
+                    echos = layout.get_echoes(subject=subj, extension=[".nii", ".nii.gz"],
                                               task=nifti_struct.entities['task'],
                                               **kwargs)
                     n_echos = len(echos)
@@ -447,7 +447,7 @@ def parse_niftis(layout, niftis, subj, config, **kwargs):
                         metadata['EchoTime'] = []
                         for echo in sorted(echos):
                             echo_struct = layout.get(subject=subj, echo=echo,
-                                                     extension='nii.gz',
+                                                     extension=[".nii", ".nii.gz"],
                                                      task=nifti_struct.entities['task'],
                                                      **kwargs)[0]
                             echo_file = echo_struct.path
@@ -469,7 +469,7 @@ def parse_niftis(layout, niftis, subj, config, **kwargs):
                 description_list.append(anat_info(suffix, metadata, img,
                                                   config))
             elif nifti_struct.entities['datatype'] == 'dwi':
-                bval_file = nii_file.replace('.nii.gz', '.bval')
+                bval_file = nii_file.replace('.nii', '.bval')
                 description_list.append(dwi_info(bval_file, metadata, img,
                                                  config))
             elif nifti_struct.entities['datatype'] == 'fmap':
