### 2019-08-05 717721f
fix renaming of .nii or .nii.gz into .bval
