### 2019-08-07 c6a88fb
adds more sensible repr to Tag; closes #466
