### 2019-08-07 b1c5728
make repr look like standard object
Co-Authored-By: Chris Markiewicz <markiewicz@stanford.edu>
