+++ bids/layout/models.py
@@ -493,7 +493,7 @@ class Tag(Base):
         self._init_on_load()
     
     def __repr__(self):
-        msg = "Tag for file '{}' with entity name '{}' has value {}."
+        msg = "<Tag file:{!r} entity:{!r} value:{!r}>"
         return msg.format(self.file_path, self.entity_name, self.value)
 
     @reconstructor
