+++ bids/layout/tests/test_models.py
@@ -101,7 +101,7 @@ def test_tag_init(sample_bidsfile, subject_entity):
     f, e = sample_bidsfile, subject_entity
     tag = Tag(f, e, 'zzz')
     rep = str(tag)
-    assert rep.startswith("Tag for file") and f.path in rep and 'zzz' in rep
+    assert rep.startswith("<Tag file:") and f.path in rep and 'zzz' in rep
 
 
 def test_tag_dtype(sample_bidsfile, subject_entity):
