+++ bids/layout/index.py
@@ -179,6 +179,15 @@ class BIDSLayoutIndexer(object):
         for c in self.config:
             all_entities.update(c.entities)
 
+        # If key/value pairs in JSON files duplicate ones extracted from files,
+        # we can end up with Tag collisions in the DB. To prevent this, we
+        # store all filename/entity name pairs in a set, and then check against
+        # it before adding each new Tag.
+
+        all_tags = ['{}_{}_{}'.format(t.file_path, t.entity_name, t.value)
+                    for t in self.session.query(Tag).all()]
+        all_tags = set(all_tags)
+
         # We build up a store of all file data as we iterate files. It looks
         # like: { extension/suffix: dirname: [(entities, payload)]}}.
         # The payload is left empty for non-JSON files.
@@ -326,6 +335,10 @@ class BIDSLayoutIndexer(object):
 
             # Create Tag <-> Entity mappings, and any newly discovered Entities
             for md_key, md_val in file_md.items():
+                tag_string = '{}_{}_{}'.format(bf.path, md_key, md_val)
+                # Skip pairs that were already found in the filenames
+                if tag_string in all_tags:
+                    continue
                 if md_key not in all_entities:
                     all_entities[md_key] = Entity(md_key, is_metadata=True)
                     self.session.add(all_entities[md_key])
