### 2019-08-07 cc5505b
raise exception in case of filename/metadata entity conflict
