### 2019-08-09 #1294 @effigies
Turns out, @tyarkoni doesn't have the magic touch for Zenodo. It's @yarikoptic who started the archive (https://github.com/bids-standard/pybids/issues/299), so only he has the power.

Yarik, would you mind updating the title of https://zenodo.org/record/3333492 (specifically that one) to "PyBIDS: Python tools for BIDS datasets".
