### 2019-08-09 #1294 @yarikoptic
- how do I share the super powers?
- updated title for that one and 0.9.3 -- clicked Save -- but do not see that reflected anywhere
