### 2019-08-10 b4b0d3b
RF: Use extras_require to declare dependencies for tests
