### 2019-08-12 #1294 @yarikoptic
> @yarikoptic I wonder if hitting "Publish" instead of "Save" would make a difference.

Sorry for the delay!!! I just did it and it published.  But now it doesn't show the version in the title.  I wonder if I should add it (`PyBIDS: Python tools for BIDS datasets (0.9.2)`) and also adjust subsequent release records?
