### 2019-08-12 #1294 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
