### 2019-08-12 #1294 @danielskatz
Thanks to @grlee77 for reviewing and to @cMadan for editing
