### 2019-08-12 #1294 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
